Refactor periscope into modules
