+++ rfmux/tools/periscope.py
@@ -3,3202 +3,14 @@
 Periscope – Real‑Time Multi‑Pane Viewer
 =======================================
 
-Features:
-- Time‑domain (TOD)
-- IQ (density or scatter)
-- FFT (pyqtgraph's fftMode=True)
-- Single‑Sideband PSD (SSB)
-- Dual‑Sideband PSD (DSB)
-- Network Analysis (amplitude and phase vs frequency)
-
-USAGE
------
-Command-Line Examples:
-- Requires rfmux to be installed as a python package
-    - `$ cd rfmux ; pip install -e .`
-- `$ periscope rfmux0022.local --module 2 --channels "3&5,7"`
-IPython / Jupyter: invoke directly from CRS object
-- `>>> crs.raise_periscope(module=2, channels="3&5")`
-- If in a non-blocking mode, you can still interact with your session concurrently.
-
-ARCHITECTURAL SUMMARY
----------------------
-This module implements the Periscope real-time multi-pane viewer using PyQt6. It
-visualizes data from a CRS data streamer in multiple ways:
-    - Time-domain (TOD)
-    - IQ (density or scatter)
-    - FFT (using pyqtgraph's fftMode=True)
-    - Single-Sideband PSD (SSB)
-    - Dual-Sideband PSD (DSB)
-    - Network Analysis (amplitude and phase vs frequency)
-
-Key Components & Concurrency:
-    - A ring buffer (Circular) for each channel stores incoming data.
-    - A UDPReceiver runs in its own QThread to receive streaming data asynchronously.
-    - Separate IQTask and PSDTask workers offload expensive computations to a QThreadPool.
-    - The main Periscope class orchestrates the UI, buffer management, and dispatch of tasks.
-    - NetworkAnalysisTask runs network analyses in the background and sends updates to the GUI.
-
-Performance Notes:
-    - GPUs are hard, and Python doesn't love concurrency, so performance is dictated
-      by the single-threaded nature of PyQtGraph's rendering engine. So any time you have
-      very long buffers being plotted it will be pokey. If you want buttery smooth performance
-      for many different channels at once, the best way to get this is to launch multiple
-      instances.
-
---------------------------------------------------------------------------------
+This module now delegates most of its implementation to helper modules
+for clarity. The original functionality remains unchanged.
 """
 
-import argparse, textwrap
-import math
-import os
-import threading
-import queue
-import socket
-import sys, warnings, ctypes.util, ctypes, platform
-import time
-import random
-import pickle
-import traceback
-import csv
-import datetime
-from typing import Dict, List, Optional, Any, Tuple
-import numpy as np
-import concurrent.futures
+from .periscope_utils import *  # constants and helper functions
+from .periscope_tasks import *  # worker threads
+from .periscope_ui import *     # dialog and window classes
 
-# Create session
-import asyncio
-from ..core.session import load_session
-from ..core.schema import CRS
-
-# ───────────────────────── Global Constants ─────────────────────────
-# Display settings
-LINE_WIDTH = 1.5
-UI_FONT_SIZE = 12
-DENSITY_GRID = 512
-DENSITY_DOT_SIZE = 1
-SMOOTH_SIGMA = 1.3
-LOG_COMPRESS = True
-SCATTER_POINTS = 1_000
-SCATTER_SIZE = 5
-
-# Network analysis defaults
-DEFAULT_MIN_FREQ = 100e6  # 100 MHz
-DEFAULT_MAX_FREQ = 2450e6  # 2.45 GHz
-DEFAULT_CABLE_LENGTH = 10  # meters
-DEFAULT_AMPLITUDE = 0.001
-DEFAULT_MAX_CHANNELS = 1023
-DEFAULT_MAX_SPAN = 500e6  # 500 MHz
-DEFAULT_NPOINTS = 5000
-DEFAULT_NSAMPLES = 10
-
-# Sampling settings
-BASE_SAMPLING = 625e6 / 256.0 / 64.0  # ≈38 147.46 Hz base for dec=0
-DEFAULT_BUFFER_SIZE = 5_000
-DEFAULT_REFRESH_MS = 33
-
-# GUI update intervals
-NETANAL_UPDATE_INTERVAL = 0.1  # seconds
-
-ICON_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'icons', 'periscope-icon.svg')
-
-# ───────────────────────── Utility Functions ─────────────────────────
-
-def _check_xcb_cursor_runtime() -> bool:
-    """
-    Detects whether libxcb-cursor is available on the current system.
-    Returns True if it can be opened, False otherwise (after emitting a warning).
-    """
-    if platform.system() != "Linux":        # macOS/Windows ship Qt plugins that do not need it
-        return True
-
-    libname = ctypes.util.find_library("xcb-cursor")   # returns None if not found
-    if libname is None:
-        warnings.warn(
-            "System library 'libxcb-cursor0' is missing.\n"
-            "Install it with:  sudo apt-get install libxcb-cursor0  (Debian/Ubuntu)\n"
-            "or the equivalent package for your distribution.",
-            RuntimeWarning,
-        )
-        return False
-    try:
-        ctypes.CDLL(libname)    # will raise OSError on broken symlink
-        return True
-    except OSError as exc:
-        warnings.warn(
-            f"libxcb-cursor was found ({libname}) but cannot be loaded: {exc}. Try reinstalling it.",
-            RuntimeWarning,
-        )
-        return False
-
-_check_xcb_cursor_runtime()
-
-# Import PyQt only after checking XCB dependencies
-from PyQt6 import QtCore, QtWidgets
-from PyQt6.QtCore import QRunnable, QThreadPool, QObject, pyqtSignal, Qt, QRegularExpression
-from PyQt6.QtGui import QFont, QIntValidator, QIcon, QDoubleValidator, QRegularExpressionValidator
-import pyqtgraph as pg
-
-# Imports for embedded iPython console
-try:
-    from qtconsole.rich_jupyter_widget import RichJupyterWidget
-    from qtconsole.inprocess import QtInProcessKernelManager
-    QTCONSOLE_AVAILABLE = True
-except ImportError:
-    QTCONSOLE_AVAILABLE = False
-    warnings.warn(
-        "qtconsole or ipykernel not found. Interactive session feature will be disabled.\n"
-        "Install them with: pip install qtconsole ipykernel",
-        RuntimeWarning,
-    )
-
-# Local imports
-import rfmux # Ensure rfmux is available for the console
-from .. import streamer
-from ..awaitless import load_ipython_extension as load_awaitless_extension
-from ..core.transferfunctions import (
-    spectrum_from_slow_tod,
-    convert_roc_to_volts,
-    convert_roc_to_dbm,
-    fit_cable_delay,
-    calculate_new_cable_length,
-    recalculate_displayed_phase,
-    EFFECTIVE_PROPAGATION_SPEED, # For direct use if needed, though functions encapsulate it
-)
-from ..algorithms.measurement import fitting # Added for find_resonances
-from ..core.hardware_map import macro
-from ..core.schema import CRS
-
-# Configure PyQtGraph
-pg.setConfigOptions(useOpenGL=False, antialias=False)
-
-# Try to import optional SciPy features
-try:
-    from scipy.ndimage import gaussian_filter, convolve
-except ImportError:  # SciPy not installed – graceful degradation
-    gaussian_filter = None
-    convolve = None
-    SMOOTH_SIGMA = 0.0
-
-def _pin_current_thread_to_core():
-    """
-    Pins the calling thread to a randomly selected CPU core from the set of
-    cores this process is allowed to run on (Linux only). On other platforms
-    or if unsupported, it emits a warning and does nothing.
-
-    Notes
-    -----
-    - Uses os.sched_getaffinity(0) to retrieve the process's allowed CPU cores.
-    - Picks one randomly (via random.choice).
-    - Calls os.sched_setaffinity(tid, {chosen_core}) for the native thread ID.
-    - Requires Python 3.8+ for threading.get_native_id().
-    - If permission is denied or the environment doesn't support thread affinity,
-      it emits a warning but continues.
-    """
-    if platform.system() != "Linux":
-        warnings.warn(
-            "Warning: Thread pinning is only supported on Linux. "
-            "Performance may suffer\n"
-        )
-        return
-    try:
-        # Retrieve the set of CPU cores the current process is allowed to use
-        allowed_cores = os.sched_getaffinity(0)
-        if not allowed_cores:
-            warnings.warn(
-                "Warning: No cores found in sched_getaffinity(0). "
-                "Not pinning.\n"
-            )
-            return
-
-        # Select one core randomly
-        chosen_core = random.choice(list(allowed_cores))
-
-        # Get the OS thread ID (Python 3.8+)
-        tid = threading.get_native_id()
-
-        # Pin this thread to the chosen core
-        os.sched_setaffinity(tid, {chosen_core})
-
-    except (AttributeError, NotImplementedError, PermissionError) as e:
-        # Possibly older Python or lacking permissions
-        warnings.warn(f"Warning: Could not pin thread to a single CPU: {e}\n")
-
-    except Exception as ex:
-        warnings.warn(f"Warning: Unexpected error pinning thread to single CPU: {ex}\n")
-
-def _get_ipython():
-    """
-    Attempt to import and return IPython's get_ipython() if available.
-
-    Returns
-    -------
-    IPython.core.interactiveshell.InteractiveShell or None
-        The interactive IPython instance, or None if IPython is not present.
-    """
-    try:
-        from IPython import get_ipython
-        return get_ipython()
-    except ImportError:
-        return None
-
-def _is_qt_event_loop_running() -> bool:
-    """
-    Check if a Qt event loop is active in IPython.
-
-    Returns
-    -------
-    bool
-        True if a Qt event loop is running, False otherwise.
-    """
-    ip = _get_ipython()
-    return bool(ip and getattr(ip, "active_eventloop", None) == "qt")
-
-def _is_running_inside_ipython() -> bool:
-    """
-    Check if the current environment is an IPython shell.
-
-    Returns
-    -------
-    bool
-        True if running in IPython, otherwise False.
-    """
-    return _get_ipython() is not None
-
-def infer_dec_stage(fs: float) -> int:
-    """
-    Estimate an integer decimation stage from a measured sample rate.
-
-    The decimation is derived by taking a log2 ratio of the base sampling
-    frequency (~38.147 kHz) to fs, rounded to the nearest integer, clamped
-    to [0..15].
-
-    Parameters
-    ----------
-    fs : float
-        Measured sample rate.
-
-    Returns
-    -------
-    int
-        Decimation stage in the range [0..15].
-    """
-    if fs < 1.0:
-        return 0
-    ratio = BASE_SAMPLING / fs
-    dec_approx = np.log2(ratio) if ratio > 0 else 0
-    dec_rounded = int(round(dec_approx))
-    return max(0, min(15, dec_rounded))
-
-def _parse_channels_multich(txt: str) -> List[List[int]]:
-    """
-    Parse a comma-separated string of channel specifications, supporting '&'
-    to group multiple channels on a single row.
-
-    Examples
-    --------
-    "3&5" => [[3,5]]
-    "3,5" => [[3],[5]]
-    "2&4,7&9" => [[2,4],[7,9]]
-    "1,2,3&8" => [[1],[2],[3,8]]
-
-    Parameters
-    ----------
-    txt : str
-        A comma-separated list of channels. Ampersand '&' merges them into one row.
-
-    Returns
-    -------
-    list of list of int
-        Each sub-list is a row containing one or more channels. If none found, returns [[1]].
-    """
-    out = []
-    for token in txt.split(","):
-        token = token.strip()
-        if not token:
-            continue
-        if "&" in token:
-            subs = token.split("&")
-            group = []
-            for sub in subs:
-                sub = sub.strip()
-                if sub.isdigit():
-                    c = int(sub)
-                    if 1 <= c <= streamer.NUM_CHANNELS:
-                        group.append(c)
-            if group:
-                out.append(group)
-        else:
-            if token.isdigit():
-                c = int(token)
-                if 1 <= c <= streamer.NUM_CHANNELS:
-                    out.append([c])
-    if not out:
-        return [[1]]
-    return out
-
-def _mode_title(mode: str) -> str:
-    """
-    Provide a more user-friendly label for each plot mode.
-
-    Parameters
-    ----------
-    mode : str
-        One of {"T", "IQ", "F", "S", "D", "NA"}.
-
-    Returns
-    -------
-    str
-        A human-readable title segment for that mode.
-    """
-    mode_titles = {
-        "T": "Time",
-        "IQ": "IQ",
-        "F": "Raw FFT",
-        "S": "SSB PSD",
-        "D": "DSB PSD",
-        "NA": "Network Analysis"
-    }
-    return mode_titles.get(mode, mode)
-
-# ───────────────────────── Unit Conversion ─────────────────────────
-class UnitConverter:
-    """
-    Utility class for converting between different units.
-    """
-    
-    @staticmethod
-    def normalize_to_dbm(normalized_amplitude: float, dac_scale_dbm: float, resistance: float = 50.0) -> float:
-        """
-        Convert normalized amplitude (0-1) to dBm power.
-        
-        Parameters
-        ----------
-        normalized_amplitude : float
-            Normalized amplitude value (0-1)
-        dac_scale_dbm : float
-            DAC scale in dBm
-        resistance : float, optional
-            Load resistance in ohms, default is 50.0
-            
-        Returns
-        -------
-        float
-            Power in dBm
-        """
-        if normalized_amplitude <= 0:
-            return float('-inf')  # -infinity dBm for zero amplitude
-            
-        # Calculate maximum voltage from dac_scale_dbm (peak voltage)
-        power_max_mw = 10**(dac_scale_dbm/10)
-        power_max_w = power_max_mw / 1000
-        v_rms_max = np.sqrt(power_max_w * resistance)
-        v_peak_max = v_rms_max * np.sqrt(2.0)
-        
-        # Calculate peak voltage from normalized amplitude
-        v_peak = normalized_amplitude * v_peak_max
-        
-        # Convert to RMS for power calculation
-        v_rms = v_peak / np.sqrt(2.0)
-        
-        # Calculate power and convert to dBm
-        power_w = v_rms**2 / resistance
-        power_mw = power_w * 1000
-        dbm = 10 * np.log10(power_mw)
-        
-        return dbm
-
-    @staticmethod
-    def dbm_to_normalize(dbm: float, dac_scale_dbm: float, resistance: float = 50.0) -> float:
-        """
-        Convert dBm power to normalized amplitude (0-1).
-        
-        Parameters
-        ----------
-        dbm : float
-            Power in dBm
-        dac_scale_dbm : float
-            DAC scale in dBm
-        resistance : float, optional
-            Load resistance in ohms, default is 50.0
-            
-        Returns
-        -------
-        float
-            Normalized amplitude (0-1)
-        """
-        # Convert dBm to RMS voltage
-        power_mw = 10**(dbm/10)
-        power_w = power_mw / 1000
-        v_rms = np.sqrt(power_w * resistance)
-        
-        # Convert RMS to peak voltage
-        v_peak = v_rms * np.sqrt(2.0)
-        
-        # Calculate maximum peak voltage from dac_scale_dbm
-        power_max_mw = 10**(dac_scale_dbm/10)
-        power_max_w = power_max_mw / 1000
-        v_rms_max = np.sqrt(power_max_w * resistance)
-        v_peak_max = v_rms_max * np.sqrt(2.0)
-        
-        # Calculate normalized amplitude
-        normalized_amplitude = v_peak / v_peak_max
-        
-        return normalized_amplitude
-
-    @staticmethod
-    def convert_amplitude(amps: np.ndarray, iq_data: np.ndarray, unit_mode: str = None, 
-                          current_mode: str = "counts", normalize: bool = False) -> np.ndarray:
-        """
-        Convert amplitude values to the specified unit, with optional normalization.
-        
-        Parameters
-        ----------
-        amps : ndarray
-            Raw amplitude values (magnitude of complex data)
-        iq_data : ndarray
-            Complex IQ data for additional conversions
-        unit_mode : str, optional
-            Unit mode to convert to: "counts", "volts", or "dbm"
-            If None, uses current_mode
-        current_mode : str, optional
-            Current unit mode if no target is specified
-        normalize : bool, optional
-            Whether to normalize the values (default: False)
-                
-        Returns
-        -------
-        ndarray
-            Converted amplitude values in the selected unit
-        """
-        # Use specified unit_mode or fall back to current_mode
-        mode = unit_mode if unit_mode is not None else current_mode
-        
-        # First convert to the right units
-        if mode == "counts":
-            result = amps.copy()  # Raw counts
-        elif mode == "volts":
-            result = convert_roc_to_volts(amps)
-        elif mode == "dbm":
-            # For network analysis, amp values are already magnitudes
-            result = convert_roc_to_dbm(amps)
-        else:
-            result = amps.copy()  # Default fallback
-            
-        # Then normalize if requested (and if there's data)
-        if normalize and len(result) > 0:
-            # For dBm, we subtract the first value instead of dividing
-            if mode == "dbm":
-                # Handle cases where the first value might be invalid (inf or nan)
-                ref_val = result[0]
-                if np.isfinite(ref_val):
-                    result = result - ref_val
-            else:
-                # For counts or volts, we divide by the first value
-                ref_val = result[0]
-                if ref_val != 0 and np.isfinite(ref_val):
-                    result = result / ref_val
-        
-        return result
-
-# ───────────────────────── Lock‑Free Ring Buffer ─────────────────────────
-class Circular:
-    """
-    A fixed-size lock-free ring buffer for time-series data.
-
-    Parameters
-    ----------
-    size : int
-        The maximum capacity of the buffer.
-    dtype : data-type, optional
-        Numpy data type for the underlying array (default float).
-    """
-
-    def __init__(self, size: int, dtype=float) -> None:
-        self.N = size
-        self.buf = np.zeros(size * 2, dtype=dtype)
-        self.ptr = 0
-        self.count = 0
-
-    def add(self, value):
-        """
-        Add a single data point to the ring buffer.
-
-        Parameters
-        ----------
-        value : scalar
-            The new data value to store.
-        """
-        self.buf[self.ptr] = value
-        self.buf[self.ptr + self.N] = value
-        self.ptr = (self.ptr + 1) % self.N
-        self.count = min(self.count + 1, self.N)
-
-    def data(self) -> np.ndarray:
-        """
-        Retrieve the valid data from the buffer as a contiguous NumPy slice.
-
-        Returns
-        -------
-        np.ndarray
-            The valid contiguous section of the buffer.
-        """
-        if self.count < self.N:
-            return self.buf[: self.count]
-        return self.buf[self.ptr : self.ptr + self.N]
-
-# ───────────────────────── Custom Plot Controls ─────────────────────────
-class ClickableViewBox(pg.ViewBox):
-    """
-    A custom ViewBox that opens a coordinate readout dialog when double-clicked
-    and supports rectangle selection zooming.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Default to RectMode (box zoom)
-        self.setMouseMode(pg.ViewBox.RectMode)  
-
-    def enableZoomBoxMode(self, enable=True):
-        """
-        Enable or disable zoom box mode.
-        
-        Parameters
-        ----------
-        enable : bool
-            If True, enables rectangle selection zooming. If False, returns to pan mode.
-        """
-        self.setMouseMode(pg.ViewBox.RectMode if enable else pg.ViewBox.PanMode)
-
-    def mouseDoubleClickEvent(self, ev):
-        """
-        Handle double-click events to display the point's X/Y coordinates
-        in a message box.
-
-        Parameters
-        ----------
-        ev : QMouseEvent
-            The mouse event instance.
-        """
-        window = getattr(self, 'parent_window', None)
-        log_x, log_y = self.state["logMode"]
-        pt_view = self.mapSceneToView(ev.scenePos())
-        x_val = 10 ** pt_view.x() if log_x else pt_view.x()
-
-        if window and getattr(window, 'add_subtract_mode', False):
-            freq = x_val
-            if ev.button() == QtCore.Qt.MouseButton.RightButton or ev.modifiers() & QtCore.Qt.KeyboardModifier.ShiftModifier:
-                window._remove_resonance(self.module_id, freq)
-            else:
-                window._add_resonance(self.module_id, freq)
-            ev.accept()
-            return
-        if ev.button() != QtCore.Qt.MouseButton.LeftButton:
-            super().mouseDoubleClickEvent(ev)
-            return
-        plot_item = self.parentItem()
-        x_label = y_label = ""
-        if isinstance(plot_item, pg.PlotItem):
-            x_axis = plot_item.getAxis("bottom")
-            y_axis = plot_item.getAxis("left")
-            if x_axis and x_axis.label:
-                x_label = x_axis.label.toPlainText().strip()
-            if y_axis and y_axis.label:
-                y_label = y_axis.label.toPlainText().strip()
-        x_label = x_label or "X"
-        y_label = y_label or "Y"
-
-        y_val = 10 ** pt_view.y() if log_y else pt_view.y()
-
-        parent_widget = self.scene().views()[0].window()
-        box = QtWidgets.QMessageBox(parent_widget)
-        box.setWindowTitle("Coordinates")
-        box.setText(f"{y_label}: {y_val:.6g}\n{x_label}: {x_val:.6g}")
-        box.setStandardButtons(QtWidgets.QMessageBox.StandardButton.Close)
-        box.setAttribute(QtCore.Qt.WidgetAttribute.WA_DeleteOnClose)
-        box.show()
-        ev.accept()
-
-# ───────────────────────── Network Data Processing ─────────────────────────
-class UDPReceiver(QtCore.QThread):
-    """
-    Receives multicast packets in a dedicated QThread and pushes them
-    into a thread-safe queue.
-
-    Parameters
-    ----------
-    host : str
-        The multicast or UDP host address.
-    module : int
-        The module number to filter on.
-    """
-
-    def __init__(self, host: str, module: int) -> None:
-        super().__init__()
-        self.module_id = module
-        self.queue = queue.Queue()
-        self.sock = streamer.get_multicast_socket(host)
-        self.sock.settimeout(0.2)
-
-    def run(self):
-        """
-        Main loop that receives packets from the socket, parses them, and
-        adds them to the internal queue. Exits upon thread interruption
-        or socket error.
-        """
-        while not self.isInterruptionRequested():
-            try:
-                data = self.sock.recv(streamer.STREAMER_LEN)
-                pkt = streamer.DfmuxPacket.from_bytes(data)
-            except socket.timeout:
-                continue
-            except OSError:
-                break
-            if pkt.module == self.module_id - 1:
-                self.queue.put(pkt)
-
-    def stop(self):
-        """
-        Signal the thread to stop, and close the socket.
-        """
-        self.requestInterruption()
-        try:
-            self.sock.close()
-        except OSError:
-            pass
-
-# ───────────────────────── IQ Task & Signals ─────────────────────────
-class IQSignals(QObject):
-    """
-    Holds custom signals emitted by IQ tasks.
-    """
-    done = pyqtSignal(int, str, object)
-    # Emitted with arguments: (row, mode_string, payload)
-    #   row: row index
-    #   mode_string: "density" or "scatter"
-    #   payload: data from the computation
-
-class IQTask(QRunnable):
-    """
-    Off-thread worker for computing IQ scatter or density histograms.
-
-    Parameters
-    ----------
-    row : int
-        Row index in the channel list (maps results back to the correct row).
-    ch : int
-        Actual channel number from which the data originates.
-    I : np.ndarray
-        Array of I samples.
-    Q : np.ndarray
-        Array of Q samples.
-    dot_px : int
-        Dot diameter in pixels for point dilation in density mode.
-    mode : {"density", "scatter"}
-        Determine whether to compute a 2D histogram or scatter points.
-    signals : IQSignals
-        An IQSignals instance for communicating results back to the GUI thread.
-    """
-
-    def __init__(self, row, ch, I, Q, dot_px, mode, signals: IQSignals):
-        super().__init__()
-        self.row = row
-        self.ch = ch
-        self.I = I.copy()
-        self.Q = Q.copy()
-        self.dot_px = dot_px
-        self.mode = mode
-        self.signals = signals
-
-    def run(self):
-        """
-        Perform the required IQ computation off the main thread. Emit results
-        via self.signals.done.
-        """
-        if len(self.I) < 2:
-            self._handle_insufficient_data()
-            return
-
-        if self.mode == "density":
-            payload = self._compute_density()
-        else:  # scatter mode
-            payload = self._compute_scatter()
-
-        self.signals.done.emit(self.row, self.mode, payload)
-        
-    def _handle_insufficient_data(self):
-        """Handle the edge case of insufficient data."""
-        if self.mode == "density":
-            empty = np.zeros((DENSITY_GRID, DENSITY_GRID), np.uint8)
-            payload = (empty, (0, 1, 0, 1))
-        else:
-            payload = ([], [], [])
-        self.signals.done.emit(self.row, self.mode, payload)
-        
-    def _compute_density(self):
-        """Compute IQ density histogram."""
-        g = DENSITY_GRID
-        hist = np.zeros((g, g), np.uint32)
-
-        Imin, Imax = self.I.min(), self.I.max()
-        Qmin, Qmax = self.Q.min(), self.Q.max()
-        if Imin == Imax or Qmin == Qmax:
-            return (hist.astype(np.uint8), (Imin, Imax, Qmin, Qmax))
-
-        # Map I/Q data to pixel indices
-        ix = ((self.I - Imin) * (g - 1) / (Imax - Imin)).astype(np.intp)
-        qy = ((self.Q - Qmin) * (g - 1) / (Qmax - Qmin)).astype(np.intp)
-
-        # Base histogram
-        np.add.at(hist, (qy, ix), 1)
-
-        # Optional dot dilation
-        if self.dot_px > 1:
-            self._apply_dot_dilation(hist, ix, qy, g)
-
-        # Optional smoothing & log-compression
-        if gaussian_filter is not None and SMOOTH_SIGMA > 0:
-            hist = gaussian_filter(hist.astype(np.float32), SMOOTH_SIGMA, mode="nearest")
-        if LOG_COMPRESS:
-            hist = np.log1p(hist, out=hist.astype(np.float32))
-
-        # 8-bit normalization
-        if hist.max() > 0:
-            hist = (hist * (255.0 / hist.max())).astype(np.uint8)
-
-        return (hist, (Imin, Imax, Qmin, Qmax))
-        
-    def _apply_dot_dilation(self, hist, ix, qy, g):
-        """Apply dot dilation to the histogram."""
-        r = self.dot_px // 2
-        if convolve is not None:
-            # Faster path if SciPy is available
-            k = 2 * r + 1
-            kernel = np.ones((k, k), dtype=np.uint8)
-            hist = convolve(hist, kernel, mode="constant", cval=0)
-        else:
-            # Fallback
-            for dy in range(-r, r + 1):
-                for dx in range(-r, r + 1):
-                    ys, xs = qy + dy, ix + dx
-                    mask = ((0 <= ys) & (ys < g) &
-                            (0 <= xs) & (xs < g))
-                    np.add.at(hist, (ys[mask], xs[mask]), 1)
-                    
-    def _compute_scatter(self):
-        """Compute IQ scatter points."""
-        N = len(self.I)
-        if N > SCATTER_POINTS:
-            idx = np.linspace(0, N - 1, SCATTER_POINTS, dtype=np.intp)
-        else:
-            idx = np.arange(N, dtype=np.intp)
-        xs, ys = self.I[idx], self.Q[idx]
-        rel = idx / (idx.max() if idx.size else 1)
-        colors = pg.colormap.get("turbo").map(
-            rel.astype(np.float32), mode="byte"
-        )
-        return (xs, ys, colors)
-
-# ───────────────────────── PSD Task & Signals ─────────────────────────
-class PSDSignals(QObject):
-    """
-    Holds custom signals emitted by PSD tasks.
-    """
-    done = pyqtSignal(int, str, int, object)
-    # Emitted with arguments: (row, mode_string, channel, payload)
-
-class PSDTask(QRunnable):
-    """
-    Off‑thread worker for single or dual sideband PSD computation.
-
-    Parameters
-    ----------
-    row : int
-        Row index in the channel list (used to map results back to the correct UI row).
-    ch : int
-        Actual channel number (data source).
-    I : np.ndarray
-        Array of I samples.
-    Q : np.ndarray
-        Array of Q samples.
-    mode : {"SSB", "DSB"}
-        Determines the type of PSD computation.
-    dec_stage : int
-        Decimation stage for the spectrum_from_slow_tod() call.
-    real_units : bool
-        If True, convert PSD to dBm/dBc. Otherwise, keep as raw counts²/Hz.
-    psd_absolute : bool
-        If True and real_units is True, uses absolute (dBm) reference. Otherwise relative (dBc).
-    segments : int
-        Number of segments for Welch segmentation. Data is split by nperseg = data_len // segments.
-    signals : PSDSignals
-        A PSDSignals instance for communicating results back to the GUI thread.
-    """
-
-    def __init__(
-        self,
-        row: int,
-        ch: int,
-        I: np.ndarray,
-        Q: np.ndarray,
-        mode: str,
-        dec_stage: int,
-        real_units: bool,
-        psd_absolute: bool,
-        segments: int,
-        signals: PSDSignals,
-    ):
-        super().__init__()
-        self.row = row
-        self.ch = ch
-        self.I = I.copy()
-        self.Q = Q.copy()
-        self.mode = mode
-        self.dec_stage = dec_stage
-        self.real_units = real_units
-        self.psd_absolute = psd_absolute
-        self.segments = segments
-        self.signals = signals
-
-    def run(self):
-        """
-        Perform PSD computations off the main thread. Emit the resulting
-        frequency array(s) and PSD(s) via signals.done.
-        """
-        data_len = len(self.I)
-        if data_len < 2:
-            self._handle_insufficient_data()
-            return
-
-        # Determine reference type
-        ref = self._get_reference_type()
-        
-        # Calculate segment size
-        nper = max(1, data_len // max(1, self.segments))
-
-        # Compute spectrum based on mode
-        if self.mode == "SSB":
-            payload = self._compute_ssb_psd(ref, nper)
-        else:  # "DSB"
-            payload = self._compute_dsb_psd(ref, nper)
-
-        self.signals.done.emit(self.row, self.mode, self.ch, payload)
-
-    def _handle_insufficient_data(self):
-        """Handle the edge case of insufficient data."""
-        if self.mode == "SSB":
-            payload = ([], [], [], [], [], [], 0.0)
-        else:
-            payload = ([], [])
-        self.signals.done.emit(self.row, self.mode, self.ch, payload)
-        
-    def _get_reference_type(self) -> str:
-        """Determine the reference type for the spectrum computation."""
-        if not self.real_units:
-            return "counts"
-        return "absolute" if self.psd_absolute else "relative"
-        
-    def _compute_ssb_psd(self, ref, nper):
-        """Compute single-sideband PSD."""
-        # First compute IQ spectrum
-        spec_iq = spectrum_from_slow_tod(
-            i_data=self.I,
-            q_data=self.Q,
-            dec_stage=self.dec_stage,
-            scaling="psd",
-            reference=ref,
-            nperseg=nper,
-            spectrum_cutoff=0.9,
-        )
-        
-        # Then compute magnitude spectrum
-        M_data = np.sqrt(self.I**2 + self.Q**2)
-        spec_m = spectrum_from_slow_tod(
-            i_data=M_data,
-            q_data=np.zeros_like(M_data),
-            dec_stage=self.dec_stage,
-            scaling="psd",
-            reference=ref,
-            nperseg=nper,
-            spectrum_cutoff=0.9,
-        )
-        
-        # Return combined results
-        return (
-            spec_iq["freq_iq"],
-            spec_iq["psd_i"],
-            spec_iq["psd_q"],
-            spec_m["psd_i"],
-            spec_m["freq_iq"],
-            spec_m["psd_i"],
-            float(self.dec_stage),
-        )
-        
-    def _compute_dsb_psd(self, ref, nper):
-        """Compute dual-sideband PSD."""
-        # Compute IQ spectrum
-        spec_iq = spectrum_from_slow_tod(
-            i_data=self.I,
-            q_data=self.Q,
-            dec_stage=self.dec_stage,
-            scaling="psd",
-            reference=ref,
-            nperseg=nper,
-            spectrum_cutoff=0.9,
-        )
-        
-        # Extract dual-sideband spectrum and sort by frequency
-        freq_dsb = spec_iq["freq_dsb"]
-        psd_dsb = spec_iq["psd_dual_sideband"]
-        order = np.argsort(freq_dsb)
-        
-        return (freq_dsb[order], psd_dsb[order])
-
-# ───────────────────────── CRS Initialization Task & Signals ─────────────────────────
-class CRSInitializeSignals(QObject):
-    """Holds custom signals emitted by CRS initialization tasks."""
-    success = pyqtSignal(str)  # success message
-    error = pyqtSignal(str)    # error message
-
-# ───────────────────────── Network Analysis ─────────────────────────
-class NetworkAnalysisSignals(QObject):
-    """
-    Holds custom signals emitted by network analysis tasks.
-    """
-    progress = pyqtSignal(int, float)  # module, progress percentage
-    data_update = pyqtSignal(int, np.ndarray, np.ndarray, np.ndarray)  # module, freqs, amps, phases
-    data_update_with_amp = pyqtSignal(int, np.ndarray, np.ndarray, np.ndarray, float)  # module, freqs, amps, phases, amplitude
-    completed = pyqtSignal(int)  # module
-    error = pyqtSignal(str)  # error message
-
-class DACScaleFetcher(QtCore.QThread):
-    """Asynchronously fetch DAC scales for all modules."""
-    dac_scales_ready = QtCore.pyqtSignal(dict)
-    
-    def __init__(self, crs):
-        super().__init__()
-        self.crs = crs
-        
-    def run(self):
-        """Fetch DAC scales for all modules."""
-        dac_scales = {}
-        # Create a new event loop for this thread
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-        
-        try:
-            self._fetch_all_dac_scales(loop, dac_scales)
-        finally:
-            loop.close()
-            
-        self.dac_scales_ready.emit(dac_scales)
-        
-    def _fetch_all_dac_scales(self, loop, dac_scales):
-        """Fetch DAC scales for all modules using the provided event loop."""
-        for module in range(1, 9):
-            try:
-                dac_scale = loop.run_until_complete(
-                    self.crs.get_dac_scale('DBM', module=module)
-                )
-                dac_scales[module] = dac_scale - 1.5  # Compensation for the balun
-            except Exception as e:
-                # Silently handle the expected "Can't access module X with low analog banking" error
-                if "Can't access module" in str(e) and "analog banking" in str(e):
-                    # Just set to None without printing error
-                    dac_scales[module] = None
-                else:
-                    # Still print unexpected ValueErrors
-                    print(f"Error fetching DAC scale for module {module}: {e}")
-                    dac_scales[module] = None
-
-class NetworkAnalysisTask(QRunnable):
-    """
-    Off-thread worker for running network analysis.
-    
-    Parameters
-    ----------
-    crs : CRS
-        CRS object from HardwareMap
-    module : int
-        Module number to run analysis on
-    params : dict
-        Network analysis parameters
-    signals : NetworkAnalysisSignals
-        Signals for communication with GUI thread
-    amplitude : float, optional
-        Specific amplitude value to use for this task
-    """
-    
-    def __init__(self, crs: "CRS", module: int, params: dict, signals: NetworkAnalysisSignals, amplitude=None):
-        super().__init__()
-        self.crs = crs
-        self.module = module
-        self.params = params
-        self.signals = signals
-        self.amplitude = amplitude if amplitude is not None else params.get('amp', DEFAULT_AMPLITUDE)
-        self._running = True
-        self._last_update_time = 0
-        self._update_interval = NETANAL_UPDATE_INTERVAL  # seconds
-        self._task = None
-        self._loop = None
-        
-    def stop(self):
-        """
-        Signal the task to stop and cancel any running asyncio tasks.
-        """
-        self._running = False
-        
-        # Cancel the async task if it exists
-        if self._task and not self._task.done() and self._loop:
-            # Schedule task cancellation in the event loop
-            self._loop.call_soon_threadsafe(self._task.cancel)
-            
-            # Clean up channels
-            if self._loop.is_running():
-                try:
-                    # Try to schedule cleanup in the event loop
-                    cleanup_future = asyncio.run_coroutine_threadsafe(
-                        self._cleanup_channels(), self._loop
-                    )
-                    # Wait for cleanup with a timeout
-                    cleanup_future.result(timeout=2.0)
-                except (asyncio.CancelledError, concurrent.futures.TimeoutError, Exception):
-                    pass
-    
-    async def _cleanup_channels(self):
-        """Clean up channels on the module being analyzed."""
-        try:
-            async with self.crs.tuber_context() as ctx:
-                # Zero out amplitudes for all channels on this module
-                for j in range(1, 1024):  # Assuming max channels is 1023
-                    ctx.set_amplitude(0, channel=j, module=self.module)
-                await ctx()
-        except Exception:
-            # Ignore errors during cleanup
-            pass
-    
-    def run(self):
-        """Execute the network analysis using the take_netanal macro."""
-        # Create event loop for async operations
-        self._loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(self._loop)
-        
-        try:
-            self._execute_network_analysis()
-        except Exception as e:
-            self.signals.error.emit(str(e))
-        finally:
-            # Clean up resources
-            self._cleanup_resources()
-            
-    def _execute_network_analysis(self):
-        """Run the network analysis operation."""
-        # Create callbacks that emit signals
-        progress_cb = self._create_progress_callback()
-        data_cb = self._create_data_callback()
-        
-        # Extract parameters
-        params = self._extract_parameters()
-
-        # Clear channels if requested
-        if params['clear_channels'] and self._running:
-            self._loop.run_until_complete(self.crs.clear_channels(module=self.module))
-
-        # Set cable length before running the analysis
-        if self._running:
-            self._loop.run_until_complete(
-                self.crs.set_cable_length(length=params['cable_length'], module=self.module)
-            )
-
-        # Create a task for the netanal operation
-        if self._running:
-            self._execute_netanal_task(params, progress_cb, data_cb)
-            
-    def _create_progress_callback(self):
-        """Create a callback for progress updates."""
-        def progress_cb(module, progress):
-            if self._running:
-                self.signals.progress.emit(module, progress)
-        return progress_cb
-        
-    def _create_data_callback(self):
-        """Create a callback for data updates."""
-        def data_cb(module, freqs_raw, amps_raw, phases_raw):
-            if self._running:
-                # Sort data by frequency before displaying
-                sort_idx = np.argsort(freqs_raw)
-                freqs = freqs_raw[sort_idx]
-                amps = amps_raw[sort_idx]
-                phases = phases_raw[sort_idx]
-                
-                # Throttle updates to prevent GUI lag
-                current_time = time.time()
-                if current_time - self._last_update_time >= self._update_interval:
-                    self._last_update_time = current_time                    
-                
-                # Emit standard update and amplitude-specific update
-                self.signals.data_update.emit(module, freqs, amps, phases)
-                self.signals.data_update_with_amp.emit(module, freqs, amps, phases, self.amplitude)
-        return data_cb
-    
-    def _extract_parameters(self):
-        """Extract parameters from the params dictionary with defaults."""
-        return {
-            'fmin': self.params.get('fmin', DEFAULT_MIN_FREQ),
-            'fmax': self.params.get('fmax', DEFAULT_MAX_FREQ),
-            'nsamps': self.params.get('nsamps', DEFAULT_NSAMPLES),
-            'npoints': self.params.get('npoints', DEFAULT_NPOINTS),
-            'max_chans': self.params.get('max_chans', DEFAULT_MAX_CHANNELS),
-            'max_span': self.params.get('max_span', DEFAULT_MAX_SPAN),
-            'cable_length': self.params.get('cable_length', DEFAULT_CABLE_LENGTH),
-            'clear_channels': self.params.get('clear_channels', True)
-        }
-        
-    def _execute_netanal_task(self, params, progress_cb, data_cb):
-        """Execute the network analysis task and handle results."""
-        netanal_coro = self.crs.take_netanal(
-            amp=self.amplitude,  # Use the specific amplitude for this task
-            fmin=params['fmin'],
-            fmax=params['fmax'],
-            nsamps=params['nsamps'],
-            npoints=params['npoints'],
-            max_chans=params['max_chans'],
-            max_span=params['max_span'],
-            module=self.module,
-            progress_callback=progress_cb,
-            data_callback=data_cb
-        )
-        
-        # Create and store the task so it can be canceled
-        self._task = self._loop.create_task(netanal_coro)
-        
-        try:
-            # Run the task and get the result
-            result = self._loop.run_until_complete(self._task)
-
-            if self._running and result:
-                # Unpack the dictionary returned by take_netanal
-                fs_sorted = result['frequencies']
-                iq_sorted = result['iq_complex']
-                phase_sorted = result['phase_degrees']
-                amp_sorted = np.abs(iq_sorted)
-
-                # Emit final data update
-                self.signals.data_update.emit(self.module, fs_sorted, amp_sorted, phase_sorted)
-                self.signals.data_update_with_amp.emit(
-                    self.module, fs_sorted, amp_sorted, phase_sorted, self.amplitude
-                )
-                self.signals.completed.emit(self.module)
-
-        except asyncio.CancelledError:
-            # Task was canceled, emit error signal
-            # No need to print traceback here as it's an expected cancellation
-            self.signals.error.emit(f"Analysis canceled for module {self.module}")
-            # Make sure to clean up channels
-            if self._loop and self._loop.is_running(): # Check if loop is still valid
-                self._loop.run_until_complete(self._cleanup_channels())
-        except KeyError as ke:
-            err_msg = f"KeyError accessing network analysis results for module {self.module}: {ke}. Expected keys 'frequencies', 'iq_complex', 'phase_degrees'."
-            print(f"ERROR: {err_msg}", file=sys.stderr)
-            traceback.print_exc(file=sys.stderr)
-            self.signals.error.emit(err_msg)
-        except Exception as e:
-            err_msg = f"Error processing network analysis results for module {self.module}: {type(e).__name__}: {e}"
-            print(f"ERROR: {err_msg}", file=sys.stderr)
-            traceback.print_exc(file=sys.stderr) # This will print the full stack trace
-            self.signals.error.emit(err_msg)
-            
-    def _cleanup_resources(self):
-        """Clean up task and event loop resources."""
-        if self._task and not self._task.done():
-            self._task.cancel()
-            
-        # Clean up the event loop
-        if self._loop and self._loop.is_running():
-            self._loop.stop()
-        if self._loop:
-            self._loop.close()
-            self._loop = None
-
-class CRSInitializeTask(QRunnable):
-    """
-    Off-thread worker for initializing the CRS board.
-    """
-    def __init__(self, crs: "CRS", module: int, irig_source: Any, clear_channels: bool, signals: CRSInitializeSignals):
-        super().__init__()
-        self.crs = crs
-        self.module = module
-        self.irig_source = irig_source
-        self.clear_channels = clear_channels
-        self.signals = signals
-        self._loop = None
-
-    def run(self):
-        """Execute the CRS initialization."""
-        self._loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(self._loop)
-        
-        try:
-            self._loop.run_until_complete(self._initialize_c_r_s())
-            self.signals.success.emit("CRS board initialized successfully.")
-        except Exception as e:
-            detailed_error = f"Error during CRS initialization: {type(e).__name__}: {str(e)}\n{traceback.format_exc()}"
-            self.signals.error.emit(detailed_error)
-        finally:
-            if self._loop and self._loop.is_running():
-                self._loop.stop()
-            if self._loop:
-                self._loop.close()
-                self._loop = None
-
-    async def _initialize_c_r_s(self):
-        """Perform the asynchronous CRS initialization steps."""
-        # It's assumed crs.TIMESTAMP_PORT.BACKPLANE, .TEST, .SMA are accessible
-        # and are the actual enum members to be passed.
-        await self.crs.set_timestamp_port(self.irig_source)
-        
-        if self.clear_channels:
-            await self.crs.clear_channels(module=self.module)
-
-# ───────────────────────── Network Analysis UI ─────────────────────────
-class NetworkAnalysisDialogBase(QtWidgets.QDialog):
-    """Base class for network analysis dialogs with shared functionality."""
-    def __init__(self, parent=None, params=None, modules=None, dac_scales=None):
-        super().__init__(parent)
-        self.params = params or {}
-        self.modules = modules or [1, 2, 3, 4]
-        self.dac_scales = dac_scales or {m: None for m in self.modules}  # Default to None (unknown)
-        self.currently_updating = False  # Flag to prevent circular updates
-        
-    def setup_amplitude_group(self, layout):
-        """Setup the amplitude settings group with normalized and dBm inputs."""
-        # Create the group box with an empty title (we'll add a title in the form layout)
-        amp_group = QtWidgets.QGroupBox()
-        amp_layout = QtWidgets.QFormLayout(amp_group)
-        
-        # Get amplitude values
-        amps = self.params.get('amps', [self.params.get('amp', DEFAULT_AMPLITUDE)])
-        amp_str = ','.join(str(a) for a in amps) if amps else str(DEFAULT_AMPLITUDE)
-        
-        # Normalized amplitude input
-        self.amp_edit = QtWidgets.QLineEdit(amp_str)
-        self.amp_edit.setToolTip("Enter a single value or comma-separated list (e.g., 0.001,0.01,0.1)")
-        amp_layout.addRow("Normalized Amplitude:", self.amp_edit)
-        
-        # dBm input 
-        self.dbm_edit = QtWidgets.QLineEdit()
-        self.dbm_edit.setToolTip("Enter a single value or comma-separated list in dBm (e.g., -30,-20,-10)")
-        amp_layout.addRow("Power (dBm):", self.dbm_edit)
-        
-        # DAC scale information
-        self.dac_scale_info = QtWidgets.QLabel("Fetching DAC scales...")
-        self.dac_scale_info.setWordWrap(True)
-        amp_layout.addRow("DAC Scale (dBm):", self.dac_scale_info)
-        
-        # Connect signals for updating between normalized and dBm
-        # Only update values during typing, but don't validate
-        self.amp_edit.textChanged.connect(self._update_dbm_from_normalized_no_validate)
-        self.dbm_edit.textChanged.connect(self._update_normalized_from_dbm_no_validate)
-        
-        # Add validation when editing is finished
-        self.amp_edit.editingFinished.connect(self._validate_normalized_values)
-        self.dbm_edit.editingFinished.connect(self._validate_dbm_values)
-        
-        # Add the group to the main layout with a title in the left column
-        layout.addRow("Amplitude Settings:", amp_group)
-        
-        return amp_group
-        
-    def _update_dbm_from_normalized_no_validate(self):
-        """Update dBm values based on normalized amplitude inputs, without validation warnings."""
-        if self.currently_updating:
-            return  # Prevent recursive updates
-            
-        # Skip if dBm field is disabled (unknown DAC scale)
-        if not self.dbm_edit.isEnabled():
-            return
-            
-        self.currently_updating = True
-        try:
-            amp_text = self.amp_edit.text().strip()
-            if not amp_text:
-                self.dbm_edit.setText("")
-                return
-                
-            # Get DAC scale for selected modules
-            dac_scale = self._get_selected_dac_scale()
-            if dac_scale is None:
-                # If scale becomes unknown, disable the field
-                self.dbm_edit.setEnabled(False)
-                self.dbm_edit.setToolTip("Unable to query current DAC scale - dBm input disabled")
-                self.dbm_edit.clear()
-                return
-                
-            # Parse normalized values
-            normalized_values = self._parse_amplitude_values(amp_text)
-            
-            # Convert to dBm
-            dbm_values = []
-            for norm in normalized_values:
-                dbm = UnitConverter.normalize_to_dbm(norm, dac_scale)
-                dbm_values.append(f"{dbm:.2f}")
-            
-            # Update dBm field
-            self.dbm_edit.setText(", ".join(dbm_values))
-        finally:
-            self.currently_updating = False
-
-    def _update_normalized_from_dbm_no_validate(self):
-        """Update normalized amplitude values based on dBm inputs, without validation warnings."""
-        if self.currently_updating:
-            return  # Prevent recursive updates
-            
-        # Skip if dBm field is disabled (unknown DAC scale)
-        if not self.dbm_edit.isEnabled():
-            return
-            
-        self.currently_updating = True
-        try:
-            dbm_text = self.dbm_edit.text().strip()
-            if not dbm_text:
-                self.amp_edit.setText("")
-                return
-                
-            # Get DAC scale for selected modules
-            dac_scale = self._get_selected_dac_scale()
-            if dac_scale is None:
-                # If scale becomes unknown, disable the field
-                self.dbm_edit.setEnabled(False)
-                self.dbm_edit.setToolTip("Unable to query current DAC scale - dBm input disabled")
-                self.dbm_edit.clear()
-                return
-                
-            # Parse dBm values
-            dbm_values = self._parse_dbm_values(dbm_text)
-            
-            # Convert to normalized amplitude
-            normalized_values = []
-            for dbm in dbm_values:
-                norm = UnitConverter.dbm_to_normalize(dbm, dac_scale)
-                normalized_values.append(f"{norm:.6f}")
-            
-            # Update normalized field
-            self.amp_edit.setText(", ".join(normalized_values))
-        finally:
-            self.currently_updating = False
-
-    def _validate_normalized_values(self):
-        """Validate normalized amplitude values when editing is finished."""
-        amp_text = self.amp_edit.text().strip()
-        if not amp_text:
-            return
-            
-        dac_scale = self._get_selected_dac_scale()
-        if dac_scale is None:
-            return
-            
-        normalized_values = self._parse_amplitude_values(amp_text)
-        
-        warnings = []
-        for norm in normalized_values:
-            if norm > 1.0:
-                warnings.append(f"Warning: Normalized amplitude {norm:.6f} > 1.0 (maximum)")
-            elif norm < 1e-4:
-                warnings.append(f"Warning: Normalized amplitude {norm:.6f} < 1e-4 (minimum recommended)")
-        
-        if warnings:
-            self._show_warning_dialog("Amplitude Warning", warnings)
-
-    def _validate_dbm_values(self):
-        """Validate dBm values when editing is finished."""
-        dbm_text = self.dbm_edit.text().strip()
-        if not dbm_text:
-            return
-            
-        dac_scale = self._get_selected_dac_scale()
-        if dac_scale is None:
-            return
-            
-        dbm_values = self._parse_dbm_values(dbm_text)
-        
-        warnings = []
-        for dbm in dbm_values:
-            if dbm > dac_scale:
-                warnings.append(f"Warning: {dbm:.2f} dBm > {dac_scale:+.2f} dBm (DAC max)")
-            
-            norm = UnitConverter.dbm_to_normalize(dbm, dac_scale)
-            if norm > 1.0:
-                warnings.append(f"Warning: {dbm:.2f} dBm gives normalized amplitude > 1.0")
-            elif norm < 1e-4:
-                warnings.append(f"Warning: {dbm:.2f} dBm gives normalized amplitude < 1e-4")
-        
-        if warnings:
-            self._show_warning_dialog("Amplitude Warning", warnings)
-
-    def _show_warning_dialog(self, title, warnings):
-        """Show a warning dialog with the given messages."""
-        QtWidgets.QMessageBox.warning(self, title, "\n".join(warnings))
-            
-    def _parse_amplitude_values(self, amp_text):
-        """Parse comma-separated amplitude values."""
-        normalized_values = []
-        for part in amp_text.split(','):
-            part = part.strip()
-            if part:
-                try:
-                    value = float(eval(part))
-                    normalized_values.append(value)
-                except (ValueError, SyntaxError, NameError):
-                    continue
-        return normalized_values
-        
-    def _parse_dbm_values(self, dbm_text):
-        """Parse comma-separated dBm values."""
-        dbm_values = []
-        for part in dbm_text.split(','):
-            part = part.strip()
-            if part:
-                try:
-                    value = float(eval(part))
-                    dbm_values.append(value)
-                except (ValueError, SyntaxError, NameError):
-                    continue
-        return dbm_values
-
-    def _update_dac_scale_info(self):
-        """Update the DAC scale information label based on selected modules."""
-        selected_modules = self._get_selected_modules()
-        
-        # Check if any selected module has a known DAC scale
-        has_known_scale = False
-        scales = []
-        
-        for m in selected_modules:
-            dac_scale = self.dac_scales.get(m)
-            if dac_scale is not None:
-                has_known_scale = True
-                scales.append(f"Module {m}: {dac_scale:+.2f} dBm")
-            else:
-                scales.append(f"Module {m}: Unknown")
-        
-        # Update the info text
-        if not selected_modules:
-            text = "Unknown (no modules selected)"
-        else:
-            text = "\n".join(scales)
-        
-        self.dac_scale_info.setText(text)
-        
-        # Enable/disable dBm input based on whether we have known scales
-        if has_known_scale:
-            self.dbm_edit.setEnabled(True)
-            self.dbm_edit.setToolTip("Enter a single value or comma-separated list in dBm (e.g., -30,-20,-10)")
-            self._update_dbm_from_normalized()  # Update conversion
-        else:
-            self.dbm_edit.setEnabled(False)
-            self.dbm_edit.setToolTip("Unable to query current DAC scale - dBm input disabled")
-            self.dbm_edit.clear()  # Clear any existing text
-    
-    def _get_selected_modules(self):
-        """
-        Get the list of currently selected modules.
-        This method should be overridden by subclasses.
-        """
-        return []
-        
-    def _get_selected_dac_scale(self):
-        """Get the DAC scale for the currently selected module(s)."""
-        selected_modules = self._get_selected_modules()
-        
-        if not selected_modules:
-            return None  # No modules selected
-            
-        # Look for the first module with a known scale
-        for module in selected_modules:
-            dac_scale = self.dac_scales.get(module)
-            if dac_scale is not None:
-                return dac_scale
-                
-        # If no selected module has a known scale, return None
-        return None
-    
-    def _update_dbm_from_normalized(self):
-        """Update dBm values based on normalized amplitude inputs."""
-        if self.currently_updating:
-            return  # Prevent recursive updates
-            
-        # Skip if dBm field is disabled (unknown DAC scale)
-        if not self.dbm_edit.isEnabled():
-            return
-            
-        self.currently_updating = True
-        try:
-            amp_text = self.amp_edit.text().strip()
-            if not amp_text:
-                self.dbm_edit.setText("")
-                return
-                
-            # Get DAC scale for selected modules
-            dac_scale = self._get_selected_dac_scale()
-            if dac_scale is None:
-                # If scale becomes unknown, disable the field
-                self.dbm_edit.setEnabled(False)
-                self.dbm_edit.setToolTip("Unable to query current DAC scale - dBm input disabled")
-                self.dbm_edit.clear()
-                return
-                
-            # Parse normalized values
-            normalized_values = self._parse_amplitude_values(amp_text)
-            
-            # Convert to dBm
-            dbm_values = []
-            warnings = []
-            
-            for norm in normalized_values:
-                if norm > 1.0:
-                    warnings.append(f"Warning: Normalized amplitude {norm:.6f} > 1.0 (maximum)")
-                elif norm < 1e-4:
-                    warnings.append(f"Warning: Normalized amplitude {norm:.6f} < 1e-4 (minimum recommended)")
-                    
-                dbm = UnitConverter.normalize_to_dbm(norm, dac_scale)
-                dbm_values.append(f"{dbm:.2f}")
-            
-            # Update dBm field
-            self.dbm_edit.setText(", ".join(dbm_values))
-            
-            # Show warnings if any
-            if warnings:
-                self._show_warning_dialog("Amplitude Warning", warnings)
-        finally:
-            self.currently_updating = False
-    
-    def _update_normalized_from_dbm(self):
-        """Update normalized amplitude values based on dBm inputs."""
-        if self.currently_updating:
-            return  # Prevent recursive updates
-            
-        # Skip if dBm field is disabled (unknown DAC scale)
-        if not self.dbm_edit.isEnabled():
-            return
-            
-        self.currently_updating = True
-        try:
-            dbm_text = self.dbm_edit.text().strip()
-            if not dbm_text:
-                self.amp_edit.setText("")
-                return
-                
-            # Get DAC scale for selected modules
-            dac_scale = self._get_selected_dac_scale()
-            if dac_scale is None:
-                # If scale becomes unknown, disable the field
-                self.dbm_edit.setEnabled(False)
-                self.dbm_edit.setToolTip("Unable to query current DAC scale - dBm input disabled")
-                self.dbm_edit.clear()
-                return
-                
-            # Parse dBm values
-            dbm_values = self._parse_dbm_values(dbm_text)
-            
-            # Convert to normalized amplitude
-            normalized_values = []
-            warnings = []
-            
-            for dbm in dbm_values:
-                if dbm > dac_scale:
-                    warnings.append(f"Warning: {dbm:.2f} dBm > {dac_scale:.2f} dBm (DAC max)")
-                
-                norm = UnitConverter.dbm_to_normalize(dbm, dac_scale)
-                
-                if norm > 1.0:
-                    warnings.append(f"Warning: {dbm:.2f} dBm gives normalized amplitude > 1.0")
-                elif norm < 1e-4:
-                    warnings.append(f"Warning: {dbm:.2f} dBm gives normalized amplitude < 1e-4")
-                    
-                normalized_values.append(f"{norm:.6f}")
-            
-            # Update normalized field
-            self.amp_edit.setText(", ".join(normalized_values))
-            
-            # Show warnings if any
-            if warnings:
-                self._show_warning_dialog("Amplitude Warning", warnings)
-        finally:
-            self.currently_updating = False
-
-class NetworkAnalysisDialog(NetworkAnalysisDialogBase):
-    """Dialog for configuring network analysis parameters with dBm support."""
-    def __init__(self, parent=None, modules=None, dac_scales=None):
-        super().__init__(parent, None, modules, dac_scales)
-        self.setWindowTitle("Network Analysis Configuration")
-        self.setModal(False)
-        self._setup_ui()
-        
-    def _setup_ui(self):
-        """Set up the user interface for the dialog."""
-        layout = QtWidgets.QVBoxLayout(self)
-        
-        # Parameters group
-        param_group = QtWidgets.QGroupBox("Analysis Parameters")
-        param_layout = QtWidgets.QFormLayout(param_group)
-        
-        # Module selection
-        self.module_entry = QtWidgets.QLineEdit("All")
-        self.module_entry.setToolTip("Enter module numbers (e.g., '1,2,5' or '1-4' or 'All')")
-        self.module_entry.textChanged.connect(self._update_dac_scale_info)
-        param_layout.addRow("Modules:", self.module_entry)
-        
-        # Frequency range (in MHz instead of Hz)
-        self.fmin_edit = QtWidgets.QLineEdit(str(DEFAULT_MIN_FREQ / 1e6))
-        self.fmax_edit = QtWidgets.QLineEdit(str(DEFAULT_MAX_FREQ / 1e6))
-        param_layout.addRow("Min Frequency (MHz):", self.fmin_edit)
-        param_layout.addRow("Max Frequency (MHz):", self.fmax_edit)
-
-        # Cable length
-        self.cable_length_edit = QtWidgets.QLineEdit(str(DEFAULT_CABLE_LENGTH))
-        param_layout.addRow("Cable Length (m):", self.cable_length_edit)        
-        
-        # Add amplitude settings group
-        self.setup_amplitude_group(param_layout)
-        
-        # Remainder of the original UI
-        self.points_edit = QtWidgets.QLineEdit(str(DEFAULT_NPOINTS))
-        param_layout.addRow("Number of Points:", self.points_edit)
-        
-        self.samples_edit = QtWidgets.QLineEdit(str(DEFAULT_NSAMPLES))
-        param_layout.addRow("Samples to Average:", self.samples_edit)
-        
-        self.max_chans_edit = QtWidgets.QLineEdit(str(DEFAULT_MAX_CHANNELS))
-        param_layout.addRow("Max Channels:", self.max_chans_edit)
-        
-        self.max_span_edit = QtWidgets.QLineEdit(str(DEFAULT_MAX_SPAN / 1e6))
-        param_layout.addRow("Max Span (MHz):", self.max_span_edit)
-        
-        self.clear_channels_cb = QtWidgets.QCheckBox("Clear all channels first")
-        self.clear_channels_cb.setChecked(True)
-        param_layout.addRow("", self.clear_channels_cb)
-        
-        layout.addWidget(param_group)
-        
-        # Buttons
-        btn_layout = QtWidgets.QHBoxLayout()
-        self.start_btn = QtWidgets.QPushButton("Start Analysis")
-        self.cancel_btn = QtWidgets.QPushButton("Cancel")
-        btn_layout.addWidget(self.start_btn)
-        btn_layout.addWidget(self.cancel_btn)
-        layout.addLayout(btn_layout)
-        
-        # Connect buttons
-        self.start_btn.clicked.connect(self.accept)
-        self.cancel_btn.clicked.connect(self.reject)
-        
-        # Initialize dBm field based on default normalized value
-        self._update_dbm_from_normalized()
-        self.setMinimumSize(500, 600)  # Width, height in pixels
-        
-    def _get_selected_modules(self):
-        """Parse module entry to determine which modules are selected."""
-        module_text = self.module_entry.text().strip()
-        selected_modules = []
-        
-        if module_text.lower() == 'all':
-            selected_modules = list(range(1, 9))  # All modules
-        else:
-            # Parse comma-separated values and ranges
-            for part in module_text.split(','):
-                part = part.strip()
-                if '-' in part:
-                    try:
-                        start, end = map(int, part.split('-'))
-                        selected_modules.extend(range(start, end + 1))
-                    except ValueError:
-                        continue
-                elif part:
-                    try:
-                        selected_modules.append(int(part))
-                    except ValueError:
-                        continue
-        
-        return selected_modules
-
-    def get_parameters(self):
-        """Get the configured parameters."""
-        try:
-            # Parse module entry
-            module_text = self.module_entry.text().strip()
-            selected_module = None
-            if module_text.lower() != 'all':
-                selected_modules = self._get_selected_modules()
-                if selected_modules:
-                    selected_module = selected_modules
-            
-            # Parse amplitude values
-            amp_text = self.amp_edit.text().strip()
-            amps = self._parse_amplitude_values(amp_text)
-            if not amps:
-                amps = [DEFAULT_AMPLITUDE]  # Default amplitude if none provided
-            
-            params = {
-                'amps': amps,
-                'module': selected_module,
-                'fmin': float(eval(self.fmin_edit.text())) * 1e6,  # Convert MHz to Hz
-                'fmax': float(eval(self.fmax_edit.text())) * 1e6,  # Convert MHz to Hz
-                'cable_length': float(self.cable_length_edit.text()),
-                'npoints': int(self.points_edit.text()),
-                'nsamps': int(self.samples_edit.text()),
-                'max_chans': int(self.max_chans_edit.text()),
-                'max_span': float(eval(self.max_span_edit.text())) * 1e6,  # Convert MHz to Hz
-                'clear_channels': self.clear_channels_cb.isChecked()
-            }
-            return params
-        except Exception as e:
-            traceback.print_exc()  # Print full stacktrace to console
-            QtWidgets.QMessageBox.critical(self, "Error", f"Invalid parameter: {str(e)}")
-            return None
-
-class NetworkAnalysisParamsDialog(NetworkAnalysisDialogBase):
-    """Dialog for editing network analysis parameters with dBm support."""
-    def __init__(self, parent=None, params=None):
-        super().__init__(parent, params)
-        self.setWindowTitle("Edit Network Analysis Parameters")
-        self.setModal(True)
-        self._setup_ui()
-        
-        if parent and hasattr(parent, 'parent') and parent.parent() is not None:
-            parent_main = parent.parent()
-            if hasattr(parent_main, 'crs') and parent_main.crs is not None:
-                self._fetch_dac_scales(parent_main.crs)
-        
-    def _fetch_dac_scales(self, crs):
-        """Fetch DAC scales for all modules."""
-        # Create a fetcher thread
-        self.fetcher = DACScaleFetcher(crs)
-        self.fetcher.dac_scales_ready.connect(self._on_dac_scales_ready)
-        self.fetcher.start()
-    
-    def _on_dac_scales_ready(self, scales):
-        """Handle fetched DAC scales."""
-        self.dac_scales = scales
-        self._update_dac_scale_info()
-        self._update_dbm_from_normalized()
-    
-    def _setup_ui(self):
-        """Set up the user interface for the dialog."""
-        layout = QtWidgets.QVBoxLayout(self)
-        
-        # Parameters form
-        form = QtWidgets.QFormLayout()
-        
-        # Frequency range (in MHz instead of Hz)
-        fmin_mhz = str(self.params.get('fmin', DEFAULT_MIN_FREQ) / 1e6)
-        fmax_mhz = str(self.params.get('fmax', DEFAULT_MAX_FREQ) / 1e6)
-        self.fmin_edit = QtWidgets.QLineEdit(fmin_mhz)
-        self.fmax_edit = QtWidgets.QLineEdit(fmax_mhz)
-        form.addRow("Min Frequency (MHz):", self.fmin_edit)
-        form.addRow("Max Frequency (MHz):", self.fmax_edit)
-        
-        # Add amplitude settings group
-        self.setup_amplitude_group(form)
-        
-        # Number of points
-        self.points_edit = QtWidgets.QLineEdit(str(self.params.get('npoints', DEFAULT_NPOINTS)))
-        form.addRow("Number of Points:", self.points_edit)
-        
-        # Number of samples to average
-        self.samples_edit = QtWidgets.QLineEdit(str(self.params.get('nsamps', DEFAULT_NSAMPLES)))
-        form.addRow("Samples to Average:", self.samples_edit)
-        
-        # Max channels
-        self.max_chans_edit = QtWidgets.QLineEdit(str(self.params.get('max_chans', DEFAULT_MAX_CHANNELS)))
-        form.addRow("Max Channels:", self.max_chans_edit)
-        
-        # Max span
-        max_span_mhz = str(self.params.get('max_span', DEFAULT_MAX_SPAN) / 1e6)
-        self.max_span_edit = QtWidgets.QLineEdit(max_span_mhz)
-        form.addRow("Max Span (MHz):", self.max_span_edit)
-        
-        # Add checkbox for clearing channels
-        self.clear_channels_cb = QtWidgets.QCheckBox("Clear all channels first")
-        self.clear_channels_cb.setChecked(self.params.get('clear_channels', True))
-        form.addRow("", self.clear_channels_cb)
-        
-        layout.addLayout(form)
-        
-        # Buttons
-        btn_layout = QtWidgets.QHBoxLayout()
-        self.ok_btn = QtWidgets.QPushButton("OK")
-        self.cancel_btn = QtWidgets.QPushButton("Cancel")
-        btn_layout.addWidget(self.ok_btn)
-        btn_layout.addWidget(self.cancel_btn)
-        layout.addLayout(btn_layout)
-        
-        # Connect buttons
-        self.ok_btn.clicked.connect(self.accept)
-        self.cancel_btn.clicked.connect(self.reject)
-        
-        # Initialize dBm field based on default normalized value
-        self._update_dbm_from_normalized()
-
-        self.setMinimumSize(500, 600)  # Width, height in pixels
-
-    def _get_selected_modules(self):
-        """Get selected modules from params."""
-        # Get selected modules from params
-        selected_module = self.params.get('module')
-        if selected_module is None:
-            # Using all modules
-            return list(range(1, 9))
-        elif isinstance(selected_module, list):
-            # Multiple specific modules
-            return selected_module
-        else:
-            # Single module
-            return [selected_module]
-    
-    def get_parameters(self):
-        """Get the updated parameters."""
-        try:
-            # Parse amplitude values
-            amp_text = self.amp_edit.text().strip()
-            amps = self._parse_amplitude_values(amp_text)
-            if not amps:
-                amps = [DEFAULT_AMPLITUDE]  # Default amplitude if none provided
-            
-            params = self.params.copy()
-            params.update({
-                'amps': amps,  # Store as list in 'amps'
-                'amp': amps[0],  # Also store first value in 'amp' for backward compatibility
-                'fmin': float(eval(self.fmin_edit.text())) * 1e6,  # Convert MHz to Hz
-                'fmax': float(eval(self.fmax_edit.text())) * 1e6,  # Convert MHz to Hz
-                'npoints': int(self.points_edit.text()),
-                'nsamps': int(self.samples_edit.text()),
-                'max_chans': int(self.max_chans_edit.text()),
-                'max_span': float(eval(self.max_span_edit.text())) * 1e6,  # Convert MHz to Hz
-                'clear_channels': self.clear_channels_cb.isChecked()
-            })
-            return params
-        except Exception as e:
-            QtWidgets.QMessageBox.critical(self, "Error", f"Invalid parameter: {str(e)}")
-            return None
-
-class NetworkAnalysisWindow(QtWidgets.QMainWindow):
-    """
-    Window for displaying network analysis results with real units support.
-    """
-    def __init__(self, parent=None, modules=None, dac_scales=None):
-        super().__init__(parent)
-        self.setWindowTitle("Network Analysis Results")
-        self.modules = modules or []
-        self.data = {}  # module -> amplitude data dictionary
-        self.raw_data = {}  # Store the raw IQ data for unit conversion
-        self.unit_mode = "dbm"  # Default to dBm instead of counts
-        self.normalize_magnitudes = False  # Add this flag to track normalization state
-        self.first_setup = True  # Flag to track initial setup
-        self.zoom_box_mode = True  # Default to zoom box mode ON
-        self.plots = {}  # Initialize plots dictionary early
-        self.original_params = {}  # Initial parameters
-        self.current_params = {}   # Most recently used parameters
-        self.dac_scales = dac_scales or {}  # Store DAC scales
-        self.resonance_lines_mag = {} # Store resonance lines for magnitude plots {module: [line_item, ...]}
-        self.resonance_lines_phase = {} # Store resonance lines for phase plots {module: [line_item, ...]}
-        self.resonance_freqs = {}  # Store resonance frequencies per module
-        self.add_subtract_mode = False
-        
-        # Setup the UI components
-        self._setup_ui()
-        # Set initial size only on creation
-        self.resize(1000, 800)
-
-    def _setup_ui(self):
-        """Set up the user interface for the window."""
-        central = QtWidgets.QWidget()
-        self.setCentralWidget(central)
-        layout = QtWidgets.QVBoxLayout(central)
-        
-        # Create toolbar
-        self._setup_toolbar(layout)
-        
-        # Create progress bars
-        self._setup_progress_bars(layout)
-        
-        # Create plot area
-        self._setup_plot_area(layout)
-
-    def _setup_toolbar(self, layout):
-        """Set up the toolbar with controls."""
-        toolbar = QtWidgets.QToolBar()
-        toolbar.setMovable(False)
-        self.addToolBar(toolbar)
-
-        # Cable length control for quick adjustments
-        self.cable_length_label = QtWidgets.QLabel("Cable Length (m):")
-        self.cable_length_spin = QtWidgets.QDoubleSpinBox()
-        self.cable_length_spin.setRange(0.0, 1000.0)
-        self.cable_length_spin.setValue(DEFAULT_CABLE_LENGTH)
-        self.cable_length_spin.setSingleStep(0.05)
-        toolbar.addWidget(self.cable_length_label)
-        toolbar.addWidget(self.cable_length_spin)
-
-        # Add Unwrap Cable Delay button
-        unwrap_button = QtWidgets.QPushButton("Unwrap Cable Delay")
-        unwrap_button.setToolTip("Fit phase slope, calculate cable length, and apply compensation.")
-        unwrap_button.clicked.connect(self._unwrap_cable_delay_action)
-        toolbar.addWidget(unwrap_button)
-
-        # Add edit parameters button to toolbar
-        edit_params_btn = QtWidgets.QPushButton("Edit Other Parameters")
-        edit_params_btn.clicked.connect(self._edit_parameters)
-        toolbar.addWidget(edit_params_btn)
-
-        # Re-run analysis button
-        rerun_btn = QtWidgets.QPushButton("Re-run Analysis")
-        rerun_btn.clicked.connect(self._rerun_analysis)
-        toolbar.addWidget(rerun_btn)
-
-        # Find Resonances button
-        find_res_btn = QtWidgets.QPushButton("Find Resonances")
-        find_res_btn.setToolTip("Identify resonance frequencies from the current sweep data.")
-        find_res_btn.clicked.connect(self._show_find_resonances_dialog)
-        toolbar.addWidget(find_res_btn)
-
-        # Show/Hide resonances checkbox
-        self.show_resonances_cb = QtWidgets.QCheckBox("Show Resonances (0)")
-        self.show_resonances_cb.setChecked(True)
-        self.show_resonances_cb.toggled.connect(self._toggle_resonances_visible)
-        toolbar.addWidget(self.show_resonances_cb)
-
-        # Add/Subtract mode and tolerance controls
-        self.edit_resonances_cb = QtWidgets.QCheckBox("Add/Subtract Resonances")
-        self.edit_resonances_cb.setToolTip(
-            "When enabled, double-click adds a resonance;\n"
-            "Shift + double-click removes the nearest resonance."
-        )
-        self.edit_resonances_cb.toggled.connect(self._toggle_resonance_edit_mode)
-        toolbar.addWidget(self.edit_resonances_cb)
-
-
-        # Export button
-        export_btn = QtWidgets.QPushButton("Export Data")
-        export_btn.clicked.connect(self._export_data)
-        toolbar.addWidget(export_btn)
-        
-        # Add spacer to push the unit controls to the far right
-        spacer = QtWidgets.QWidget()
-        spacer.setSizePolicy(QtWidgets.QSizePolicy.Policy.Expanding, 
-                            QtWidgets.QSizePolicy.Policy.Preferred)
-        toolbar.addWidget(spacer)
-        
-        # Normalize Magnitudes checkbox
-        self.normalize_checkbox = QtWidgets.QCheckBox("Normalize Magnitudes")
-        self.normalize_checkbox.setChecked(False)
-        self.normalize_checkbox.setToolTip("Normalize all magnitude curves to their first data point")
-        self.normalize_checkbox.toggled.connect(self._toggle_normalization)
-        toolbar.addWidget(self.normalize_checkbox)
-
-        # Add unit controls
-        self._setup_unit_controls(toolbar)
-        
-        # Add zoom box mode checkbox
-        self._setup_zoom_box_control(toolbar)
-
-    def _setup_unit_controls(self, toolbar):
-        """Set up the unit selection controls."""
-        unit_group = QtWidgets.QWidget()
-        unit_layout = QtWidgets.QHBoxLayout(unit_group)
-        unit_layout.setContentsMargins(0, 0, 0, 0)
-        unit_layout.setAlignment(Qt.AlignmentFlag.AlignRight)
-        
-        self.rb_counts = QtWidgets.QRadioButton("Counts")
-        self.rb_dbm = QtWidgets.QRadioButton("dBm")
-        self.rb_volts = QtWidgets.QRadioButton("Volts")
-        self.rb_dbm.setChecked(True)  # Changed from rb_counts to rb_dbm
-        
-        unit_layout.addWidget(QtWidgets.QLabel("Units:"))
-        unit_layout.addWidget(self.rb_counts)
-        unit_layout.addWidget(self.rb_dbm)
-        unit_layout.addWidget(self.rb_volts)
-        
-        # Connect signals
-        self.rb_counts.toggled.connect(lambda: self._update_unit_mode("counts"))
-        self.rb_dbm.toggled.connect(lambda: self._update_unit_mode("dbm"))
-        self.rb_volts.toggled.connect(lambda: self._update_unit_mode("volts"))
-        
-        # Set fixed size policy to make alignment more predictable
-        unit_group.setSizePolicy(QtWidgets.QSizePolicy.Policy.Fixed, 
-                                QtWidgets.QSizePolicy.Policy.Preferred)
-        
-        # Now add the unit controls at the far right
-        toolbar.addSeparator()
-        toolbar.addWidget(unit_group)
-        
-    def _setup_zoom_box_control(self, toolbar):
-        """Set up the zoom box mode control."""
-        zoom_box_cb = QtWidgets.QCheckBox("Zoom Box Mode")
-        zoom_box_cb.setChecked(self.zoom_box_mode)  # Default to ON
-        zoom_box_cb.setToolTip("When enabled, left-click drag creates a zoom box. When disabled, left-click drag pans.")
-        zoom_box_cb.toggled.connect(self._toggle_zoom_box)
-        
-        # Store reference to the checkbox
-        self.zoom_box_cb = zoom_box_cb
-        
-        toolbar.addWidget(zoom_box_cb)
-
-    def _setup_progress_bars(self, layout):
-        """Set up progress bars for each module."""
-        self.progress_group = None
-        if self.modules:
-            self.progress_group = QtWidgets.QGroupBox("Analysis Progress")
-            progress_layout = QtWidgets.QVBoxLayout(self.progress_group)
-            
-            self.progress_bars = {}
-            self.progress_labels = {}  # Add labels to show amplitude progress
-            for module in self.modules:
-                vlayout = QtWidgets.QVBoxLayout()
-                
-                # Main progress layout
-                hlayout = QtWidgets.QHBoxLayout()
-                label = QtWidgets.QLabel(f"Module {module}:")
-                pbar = QtWidgets.QProgressBar()
-                pbar.setRange(0, 100)
-                pbar.setValue(0)
-                hlayout.addWidget(label)
-                hlayout.addWidget(pbar)
-                vlayout.addLayout(hlayout)
-                
-                # Amplitude progress label
-                amp_label = QtWidgets.QLabel("")
-                amp_label.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
-                vlayout.addWidget(amp_label)
-                
-                progress_layout.addLayout(vlayout)
-                self.progress_bars[module] = pbar
-                self.progress_labels[module] = amp_label
-                
-            layout.addWidget(self.progress_group)
-        else:
-            self.progress_bars = {}
-            self.progress_labels = {}
-
-    def _setup_plot_area(self, layout):
-        """Set up the plot area with tabs for each module."""
-        self.tabs = QtWidgets.QTabWidget()
-        layout.addWidget(self.tabs)
-        
-        self.plots = {}
-        for module in self.modules:
-            tab = QtWidgets.QWidget()
-            tab_layout = QtWidgets.QVBoxLayout(tab)
-
-            # Create amplitude and phase plots with ClickableViewBox
-            vb_amp = ClickableViewBox()
-            vb_amp.parent_window = self
-            vb_amp.module_id = module
-            vb_amp.plot_role = 'amp'
-            amp_plot = pg.PlotWidget(viewBox=vb_amp, title=f"Module {module} - Magnitude")
-            self._update_amplitude_labels(amp_plot)
-            amp_plot.setLabel('bottom', 'Frequency', units='Hz')
-            amp_plot.showGrid(x=True, y=True, alpha=0.3)
-
-            vb_phase = ClickableViewBox()
-            vb_phase.parent_window = self
-            vb_phase.module_id = module
-            vb_phase.plot_role = 'phase'
-            phase_plot = pg.PlotWidget(viewBox=vb_phase, title=f"Module {module} - Phase")
-            phase_plot.setLabel('left', 'Phase', units='deg')
-            phase_plot.setLabel('bottom', 'Frequency', units='Hz')
-            phase_plot.showGrid(x=True, y=True, alpha=0.3)
-            
-            # Add legends for multiple amplitude plots
-            amp_legend = amp_plot.addLegend(offset=(30, 10))
-            phase_legend = phase_plot.addLegend(offset=(30, 10))
-
-            # Create curves with periscope color scheme - but don't add data yet
-            amp_curve = amp_plot.plot([], [], pen=pg.mkPen('#ff7f0e', width=LINE_WIDTH))  # Empty data
-            phase_curve = phase_plot.plot([], [], pen=pg.mkPen('#1f77b4', width=LINE_WIDTH))  # Empty data
-
-            tab_layout.addWidget(amp_plot)
-            tab_layout.addWidget(phase_plot)
-            self.tabs.addTab(tab, f"Module {module}")
-            
-            self.plots[module] = {
-                'amp_plot': amp_plot,
-                'phase_plot': phase_plot,
-                'amp_curve': amp_curve,
-                'phase_curve': phase_curve,
-                'amp_legend': amp_legend,
-                'phase_legend': phase_legend,
-                'amp_curves': {},  # Will store multiple curves for different amplitudes
-                'phase_curves': {},  # Will store multiple curves for different amplitudes
-                'resonance_lines_mag': [], # For storing magnitude resonance lines
-                'resonance_lines_phase': [] # For storing phase resonance lines
-            }
-            
-            # Apply zoom box mode
-            self._apply_zoom_box_mode()            
-
-            # Link the x-axis of amplitude and phase plots for synchronized zooming
-            phase_plot.setXLink(amp_plot)
-
-    def clear_plots(self):
-        """Clear all plots, curves, and legends."""
-        for module in self.plots:
-            plot_info = self.plots[module]
-            amp_plot = plot_info['amp_plot']
-            phase_plot = plot_info['phase_plot']
-            
-            # Clear legends
-            plot_info['amp_legend'].clear()
-            plot_info['phase_legend'].clear()
-            
-            # Remove all amplitude-specific curves from plots
-            for amp, curve in list(plot_info['amp_curves'].items()):
-                amp_plot.removeItem(curve)
-            for amp, curve in list(plot_info['phase_curves'].items()):
-                phase_plot.removeItem(curve)
-            
-            # Clear curve dictionaries
-            plot_info['amp_curves'].clear()
-            plot_info['phase_curves'].clear()
-            
-            # Make sure main curves have no data 
-            plot_info['amp_curve'].setData([], [])
-            plot_info['phase_curve'].setData([], [])
-
-    def update_amplitude_progress(self, module: int, current_amp: int, total_amps: int, amplitude: float):
-        """Update the amplitude progress display for a module."""
-        if hasattr(self, 'progress_labels') and module in self.progress_labels:
-            self.progress_labels[module].setText(f"Amplitude {current_amp}/{total_amps} ({amplitude})")
-
-    def _toggle_normalization(self, checked):
-        """Toggle normalization of magnitude plots."""
-        self.normalize_magnitudes = checked
-        
-        # Update axis labels
-        for module in self.plots:
-            self._update_amplitude_labels(self.plots[module]['amp_plot'])
-        
-        # Make sure main curves are cleared if we have amplitude-specific curves
-        for module in self.plots:
-            if len(self.plots[module]['amp_curves']) > 0:
-                # Clear main curves if we have amplitude-specific curves
-                self.plots[module]['amp_curve'].setData([], [])
-                self.plots[module]['phase_curve'].setData([], [])
-        
-        # Redraw plots with normalization applied
-        self._redraw_all_plots()
-
-    def _toggle_zoom_box(self, enable):
-        """Toggle zoom box mode for all plots."""
-        self.zoom_box_mode = enable
-        self._apply_zoom_box_mode()
-        
-    def _apply_zoom_box_mode(self):
-        """Apply the current zoom box mode setting to all plots."""
-        for module in self.plots:
-            for plot_type in ['amp_plot', 'phase_plot']:
-                viewbox = self.plots[module][plot_type].getViewBox()
-                if isinstance(viewbox, ClickableViewBox):
-                    viewbox.enableZoomBoxMode(self.zoom_box_mode)
-
-    def _toggle_resonances_visible(self, checked: bool):
-        """Show or hide all resonance indicator lines."""
-        for module in self.plots:
-            for line in self.plots[module]['resonance_lines_mag']:
-                line.setVisible(checked)
-            for line in self.plots[module]['resonance_lines_phase']:
-                line.setVisible(checked)
-
-    def _toggle_resonance_edit_mode(self, checked: bool):
-        """Enable or disable double-click add/subtract mode."""
-        self.add_subtract_mode = checked
-
-    def _update_resonance_checkbox_text(self, module: int):
-        """Update the show-resonances checkbox label with count."""
-        count = len(self.resonance_freqs.get(module, []))
-        self.show_resonances_cb.setText(f"Show Resonances ({count})")
-
-    def _add_resonance(self, module: int, freq_hz: float):
-        """Add a resonance line at the given frequency."""
-        if module not in self.plots:
-            return
-        plot_info = self.plots[module]
-        line_pen = pg.mkPen('r', style=QtCore.Qt.PenStyle.DashLine)
-        line_mag = pg.InfiniteLine(pos=freq_hz, angle=90, movable=False, pen=line_pen)
-        line_phase = pg.InfiniteLine(pos=freq_hz, angle=90, movable=False, pen=line_pen)
-        plot_info['amp_plot'].addItem(line_mag)
-        plot_info['phase_plot'].addItem(line_phase)
-        plot_info['resonance_lines_mag'].append(line_mag)
-        plot_info['resonance_lines_phase'].append(line_phase)
-        self.resonance_freqs.setdefault(module, []).append(freq_hz)
-        self._update_resonance_checkbox_text(module)
-        self._toggle_resonances_visible(self.show_resonances_cb.isChecked())
-
-    def _remove_resonance(self, module: int, freq_hz: float):
-        """Remove the nearest resonance line."""
-        if module not in self.plots or module not in self.resonance_freqs:
-            return
-        freqs = self.resonance_freqs[module]
-        if not freqs:
-            return
-        freqs_arr = np.array(freqs)
-        idx = int(np.argmin(np.abs(freqs_arr - freq_hz)))
-        # Remove lines
-        line_mag = self.plots[module]['resonance_lines_mag'].pop(idx)
-        line_phase = self.plots[module]['resonance_lines_phase'].pop(idx)
-        self.plots[module]['amp_plot'].removeItem(line_mag)
-        self.plots[module]['phase_plot'].removeItem(line_phase)
-        freqs.pop(idx)
-        self._update_resonance_checkbox_text(module)
-
-    def _update_unit_mode(self, mode):
-        """Update unit mode and redraw plots."""
-        if mode != self.unit_mode:
-            self.unit_mode = mode
-            # Update all plot labels
-            for module in self.plots:
-                self._update_amplitude_labels(self.plots[module]['amp_plot'])
-            
-            # Redraw with new units
-            self._redraw_all_plots()
-
-    def _update_legends_for_unit_mode(self):
-        """Update the legend entries to reflect the current unit mode."""
-        for module in self.plots:
-            # Clear existing legends
-            self.plots[module]['amp_legend'].clear()
-            self.plots[module]['phase_legend'].clear()
-            
-            # Re-add curves with updated labels
-            for amplitude, curve in self.plots[module]['amp_curves'].items():
-                # Format amplitude according to current unit mode
-                if self.unit_mode == "dbm":
-                    # Check if we have a DAC scale for this module
-                    if module not in self.dac_scales:
-                        # Issue warning and switch to counts mode
-                        print(f"Warning: No DAC scale available for module {module}, cannot display accurate probe power in dBm.")
-                        self.rb_counts.setChecked(True)  # Switch to counts mode
-                        return  # Exit and let _update_unit_mode call us again
-                    
-                    # Use the actual DAC scale without fallback
-                    dac_scale = self.dac_scales[module]
-                    dbm_value = UnitConverter.normalize_to_dbm(amplitude, dac_scale)
-                    label = f"Probe: {dbm_value:.2f} dBm"
-                elif self.unit_mode == "volts":
-                    # Properly convert normalized amplitude to voltage through power calculation
-                    # First get dBm value
-                    if module not in self.dac_scales:
-                        print(f"Warning: No DAC scale available for module {module}, cannot display accurate probe amplitude in Volts.")
-                        self.rb_counts.setChecked(True)  # Switch to counts mode
-                        return  # Exit and let _update_unit_mode call us again
-                    else:
-                        dac_scale = self.dac_scales[module]
-                        dbm_value = UnitConverter.normalize_to_dbm(amplitude, dac_scale)
-                        
-                        # Convert dBm to watts: P = 10^((dBm - 30)/10)
-                        power_watts = 10**((dbm_value - 30)/10)
-                        
-                        # Convert watts to peak voltage: V = sqrt(P * R)
-                        resistance = 50.0  # Ohms
-                        voltage_rms = np.sqrt(power_watts * resistance)
-                        
-                        # Convert RMS to peak voltage
-                        voltage_peak = voltage_rms * np.sqrt(2)
-                        voltage_peak = voltage_peak*1e6
-                        
-                        label = f"Probe: {voltage_peak:.1f} uV (peak)"
-                else:  # "counts"
-                    label = f"Probe: {amplitude} Normalized Units"
-                
-                # Add to legend with new label
-                self.plots[module]['amp_legend'].addItem(curve, label)
-                
-                # Also update the phase legend for consistency
-                phase_curve = self.plots[module]['phase_curves'].get(amplitude)
-                if phase_curve:
-                    self.plots[module]['phase_legend'].addItem(phase_curve, label)         
-    
-    def _update_amplitude_labels(self, plot):
-        """Update plot labels based on current unit mode and normalization state."""
-        if self.normalize_magnitudes:
-            if self.unit_mode == "dbm":
-                plot.setLabel('left', 'Normalized Power', units='dB') # Corrected units to dB
-            else:
-                plot.setLabel('left', 'Normalized Magnitude', units='')
-        else:
-            if self.unit_mode == "counts":
-                plot.setLabel('left', 'Magnitude', units='Counts')
-            elif self.unit_mode == "dbm":
-                plot.setLabel('left', 'Power', units='dBm')
-            elif self.unit_mode == "volts":
-                plot.setLabel('left', 'Magnitude', units='V')
-
-    def _redraw_all_plots(self):
-        """Redraw all plots with current unit mode."""
-        for module in self.raw_data:
-            if module in self.plots:
-                # Check if we have amplitude-specific curves
-                has_amp_curves = len(self.plots[module]['amp_curves']) > 0
-                
-                # Update amplitude-specific curves first
-                for amp_key, data_tuple in self.raw_data[module].items():
-                    if amp_key != 'default':
-                        # Extract amplitude and data
-                        amplitude, freqs, amps, phases, iq_data = self._extract_data_from_tuple(amp_key, data_tuple)
-                        
-                        # Update the curve if it exists
-                        if amplitude in self.plots[module]['amp_curves']:
-                            converted_amps = UnitConverter.convert_amplitude(
-                                amps, iq_data, self.unit_mode, normalize=self.normalize_magnitudes)
-                            freq_ghz = freqs / 1e9
-                            self.plots[module]['amp_curves'][amplitude].setData(freq_ghz, converted_amps)
-                            self.plots[module]['phase_curves'][amplitude].setData(freq_ghz, phases)
-                
-                # Now handle the default curve - ONLY if there are no amplitude-specific curves
-                if 'default' in self.raw_data[module] and not has_amp_curves:
-                    freqs, amps, phases, iq_data = self.raw_data[module]['default']
-                    converted_amps = UnitConverter.convert_amplitude(
-                        amps, iq_data, self.unit_mode, normalize=self.normalize_magnitudes)
-                    freq_ghz = freqs / 1e9
-                    self.plots[module]['amp_curve'].setData(freq_ghz, converted_amps)
-                    self.plots[module]['phase_curve'].setData(freq_ghz, phases)
-                else:
-                    # Make sure default curves have no data if we have amplitude-specific curves
-                    self.plots[module]['amp_curve'].setData([], [])
-                    self.plots[module]['phase_curve'].setData([], [])
-                
-                # Enable auto range to fit new data
-                self.plots[module]['amp_plot'].autoRange()
-        
-        # Update legends after redrawing curves
-        self._update_legends_for_unit_mode()              
-    
-    def _extract_data_from_tuple(self, amp_key, data_tuple):
-        """Extract amplitude and data from a data tuple."""
-        if len(data_tuple) == 5:
-            # New format with amplitude included
-            freqs, amps, phases, iq_data, amplitude = data_tuple
-        else:
-            # Old format, extract amplitude from key
-            freqs, amps, phases, iq_data = data_tuple
-            try:
-                amplitude = float(amp_key.split('_')[-1])
-            except (ValueError, IndexError):
-                amplitude = DEFAULT_AMPLITUDE
-        return amplitude, freqs, amps, phases, iq_data
-
-    def closeEvent(self, event):
-        """Handle window close event by cleaning up resources."""
-        parent = self.parent()
-        
-        if parent and hasattr(parent, 'netanal_windows'):
-            # Find our window ID
-            window_id = None
-            for w_id, w_data in parent.netanal_windows.items():
-                if w_data['window'] == self:
-                    window_id = w_id
-                    break
-            
-            if window_id:
-                # Stop all tasks for this window
-                if hasattr(parent, 'netanal_tasks'):
-                    for task_key in list(parent.netanal_tasks.keys()):
-                        if task_key.startswith(f"{window_id}_"):
-                            task = parent.netanal_tasks.pop(task_key)
-                            task.stop()
-                
-                # Remove from windows dictionary
-                parent.netanal_windows.pop(window_id, None)
-        
-        # Call parent implementation
-        super().closeEvent(event)
-
-    def _show_find_resonances_dialog(self):
-        """Show the dialog to configure and run find_resonances."""
-        current_tab_index = self.tabs.currentIndex()
-        if current_tab_index < 0:
-            QtWidgets.QMessageBox.warning(self, "No Module Selected", "Please select a module tab to analyze.")
-            return
-        
-        active_module_text = self.tabs.tabText(current_tab_index)
-        try:
-            active_module = int(active_module_text.split(" ")[1])
-        except (IndexError, ValueError):
-            QtWidgets.QMessageBox.critical(self, "Error", f"Could not determine active module from tab: {active_module_text}")
-            return
-
-        if not self.raw_data or active_module not in self.raw_data or not self.raw_data[active_module]:
-            QtWidgets.QMessageBox.information(self, "No Data", f"No sweep data available for Module {active_module} to find resonances.")
-            return
-
-        dialog = FindResonancesDialog(self)
-        if dialog.exec():
-            params = dialog.get_parameters()
-            if params:
-                self._run_and_plot_resonances(active_module, params)
-
-    def _run_and_plot_resonances(self, active_module: int, find_resonances_params: dict):
-        """Run find_resonances and plot the results on the active module's plots."""
-        module_sweeps = self.raw_data.get(active_module)
-        if not module_sweeps:
-            QtWidgets.QMessageBox.warning(self, "No Data", f"No data found for module {active_module}.")
-            return
-
-        target_sweep_key = None
-        # Try to find the last run amplitude sweep
-        ordered_amplitudes_run = self.original_params.get('amps', [])
-        if ordered_amplitudes_run:
-            for amp_setting in reversed(ordered_amplitudes_run):
-                sweep_key = f"{active_module}_{amp_setting}"
-                if sweep_key in module_sweeps:
-                    target_sweep_key = sweep_key
-                    break
-        
-        # Fallback to 'default' if no amplitude-specific sweep found or if 'amps' was empty
-        if target_sweep_key is None and 'default' in module_sweeps:
-            target_sweep_key = 'default'
-
-        if target_sweep_key is None:
-            # Fallback to the very last key added to the dictionary if all else fails
-            if module_sweeps:
-                target_sweep_key = list(module_sweeps.keys())[-1]
-
-
-        if target_sweep_key is None:
-            QtWidgets.QMessageBox.warning(self, "No Data", f"Could not determine which sweep to analyze for module {active_module}.")
-            return
-
-        # Extract data from the chosen sweep
-        data_tuple = module_sweeps[target_sweep_key]
-        
-        if len(data_tuple) == 5: # New format with amplitude
-            frequencies, _, _, iq_complex, _ = data_tuple
-        elif len(data_tuple) == 4: # Old format or default
-            frequencies, _, _, iq_complex = data_tuple
-        else:
-            QtWidgets.QMessageBox.critical(self, "Data Error", "Unexpected data format for the selected sweep.")
-            return
-
-        if len(frequencies) == 0 or len(iq_complex) == 0:
-            QtWidgets.QMessageBox.information(self, "No Data", "Selected sweep has no frequency or IQ data.")
-            return
-            
-        # Run find_resonances
-        try:
-            resonance_results = fitting.find_resonances(
-                frequencies=frequencies,
-                iq_complex=iq_complex,
-                module_identifier=f"Module {active_module} (Sweep: {target_sweep_key})",
-                **find_resonances_params
-            )
-        except Exception as e:
-            QtWidgets.QMessageBox.critical(self, "Resonance Finding Error", f"Error calling find_resonances: {str(e)}")
-            traceback.print_exc()
-            return
-
-        # Plot the results
-        if active_module in self.plots:
-            plot_info = self.plots[active_module]
-            amp_plot_item = plot_info['amp_plot'].getPlotItem()
-            phase_plot_item = plot_info['phase_plot'].getPlotItem()
-            amp_legend = plot_info['amp_legend']
-            phase_legend = plot_info['phase_legend']
-
-            # 1. Clear ALL previous resonance lines from plots and stored lists
-            for line in plot_info.get('resonance_lines_mag', []):
-                amp_plot_item.removeItem(line)
-            plot_info['resonance_lines_mag'] = []
-
-            for line in plot_info.get('resonance_lines_phase', []):
-                phase_plot_item.removeItem(line)
-            plot_info['resonance_lines_phase'] = []
-
-            # 2. Clear old resonance frequency data
-            self.resonance_freqs[active_module] = []
-            
-            res_freqs_hz = resonance_results.get('resonance_frequencies', [])
-
-            if not res_freqs_hz:
-                QtWidgets.QMessageBox.information(self, "No Resonances Found", 
-                                                  f"No resonances were identified for Module {active_module} with the given parameters.")
-                return # Important to return if no resonances, so no legend item is added
-
-            # 3. Add new lines (all initially visible)
-            line_pen = pg.mkPen('r', style=QtCore.Qt.PenStyle.DashLine)
-            for res_freq_hz in res_freqs_hz:
-                line_mag = pg.InfiniteLine(pos=res_freq_hz, angle=90, movable=False, pen=line_pen)
-                amp_plot_item.addItem(line_mag)
-                plot_info['resonance_lines_mag'].append(line_mag)
-
-                line_phase = pg.InfiniteLine(pos=res_freq_hz, angle=90, movable=False, pen=line_pen)
-                phase_plot_item.addItem(line_phase)
-                plot_info['resonance_lines_phase'].append(line_phase)
-
-            # 4. Store resonance frequencies and update checkbox
-            self.resonance_freqs[active_module] = res_freqs_hz
-            self._update_resonance_checkbox_text(active_module)
-            self._toggle_resonances_visible(self.show_resonances_cb.isChecked())
-
-
-    def _check_all_complete(self):
-        """
-        Check if all progress bars are at 100% and hide the progress group 
-        when all analyses are complete.
-        """
-        if not self.progress_group:
-            return
-            
-        # Find our window data in the parent
-        parent = self.parent()
-        window_id = None
-        
-        if parent and hasattr(parent, 'netanal_windows'):
-            for w_id, w_data in parent.netanal_windows.items():
-                if w_data['window'] == self:
-                    window_id = w_id
-                    break
-        
-        if not window_id:
-            return
-            
-        window_data = parent.netanal_windows[window_id]
-        
-        # Check if there are any pending amplitudes
-        no_pending_amplitudes = True
-        for module in window_data['amplitude_queues']:
-            if window_data['amplitude_queues'][module]:
-                no_pending_amplitudes = False
-                break
-        
-        # Hide progress when all bars are at 100% and no pending amplitudes
-        all_complete = all(pbar.value() == 100 for pbar in self.progress_bars.values())
-        if all_complete and no_pending_amplitudes:
-            self.progress_group.setVisible(False)
-
-    def _edit_parameters(self):
-        """Open dialog to edit parameters besides cable length."""
-        dialog = NetworkAnalysisParamsDialog(self, self.current_params)
-        if dialog.exec():
-            # Get updated parameters
-            params = dialog.get_parameters()
-            if params:
-                # Keep the current cable length
-                params['cable_length'] = self.cable_length_spin.value()
-                # Update current parameters
-                self.current_params = params.copy()
-                # Make progress group visible again
-                if self.progress_group:
-                    self.progress_group.setVisible(True)
-                self.parent()._rerun_network_analysis(params)
-
-    def _rerun_analysis(self):
-        """Re-run the analysis with potentially updated parameters."""
-        if hasattr(self.parent(), '_rerun_network_analysis'):
-            # Get current parameters and update cable length
-            params = self.current_params.copy()
-            params['cable_length'] = self.cable_length_spin.value()
-            # Make progress group visible again
-            if self.progress_group:
-                self.progress_group.setVisible(True)
-            self.parent()._rerun_network_analysis(params)
-    
-    def set_params(self, params):
-        """Set parameters for analysis."""
-        self.original_params = params.copy()  # Keep original for reference
-        self.current_params = params.copy()   # Keep current for dialog
-        
-        # Set cable length spinner to match value
-        self.cable_length_spin.setValue(params.get('cable_length', DEFAULT_CABLE_LENGTH))
-        
-        # Only try to set plot ranges if plots exist
-        if not hasattr(self, 'plots') or not self.plots:
-            return
-            
-        # Set initial plot ranges based on frequency parameters
-        fmin = params.get('fmin', DEFAULT_MIN_FREQ)
-        fmax = params.get('fmax', DEFAULT_MAX_FREQ)
-        for module in self.plots:
-            self.plots[module]['amp_plot'].setXRange(fmin, fmax)
-            self.plots[module]['phase_plot'].setXRange(fmin, fmax)
-            # Disable auto range on X axis but keep Y auto range
-            self.plots[module]['amp_plot'].enableAutoRange(pg.ViewBox.XAxis, False)
-            self.plots[module]['phase_plot'].enableAutoRange(pg.ViewBox.XAxis, False)
-            self.plots[module]['amp_plot'].enableAutoRange(pg.ViewBox.YAxis, True)
-            self.plots[module]['phase_plot'].enableAutoRange(pg.ViewBox.YAxis, True)
-    
-    def update_data_with_amp(self, module: int, freqs: np.ndarray, amps: np.ndarray, phases: np.ndarray, amplitude: float):
-        """Update the plot data for a specific module and amplitude."""
-        # Store the raw data for unit conversion
-        iq_data = amps * np.exp(1j * np.radians(phases))  # Reconstruct complex data
-        
-        # Use a unique key that includes the amplitude
-        key = f"{module}_{amplitude}"
-        
-        # Initialize dictionaries if needed
-        if module not in self.raw_data:
-            self.raw_data[module] = {}
-        if module not in self.data:
-            self.data[module] = {}
-        
-        # Store the data with amplitude
-        self.raw_data[module][key] = (freqs, amps, phases, iq_data, amplitude)
-        self.data[module][key] = (freqs, amps, phases)
-        
-        if module in self.plots:
-            # Hide the main curve if this is the first amplitude-specific curve
-            if len(self.plots[module]['amp_curves']) == 0:
-                # Set main curves to empty data to effectively hide them
-                self.plots[module]['amp_curve'].setData([], [])
-                self.plots[module]['phase_curve'].setData([], [])
-            
-            # Convert amplitude to selected units
-            converted_amps = UnitConverter.convert_amplitude(
-                amps, iq_data, self.unit_mode, normalize=self.normalize_magnitudes)
-            
-            # Generate a color based on the amplitude index in the list of amplitudes
-            amps_list = self.original_params.get('amps', [amplitude])
-            if amplitude in amps_list:
-                amp_index = amps_list.index(amplitude)
-            else:
-                amp_index = 0
-                
-            # Use the same color families as the main application
-            channel_families = [
-                "#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd",
-                "#8c564b", "#e377c2", "#7f7f7f", "#bcbd22", "#17becf"
-            ]
-            color = channel_families[amp_index % len(channel_families)]
-            
-            # Create or update curves for this amplitude
-            is_new_curve = False
-            if amplitude not in self.plots[module]['amp_curves']:
-                is_new_curve = True
-                self.plots[module]['amp_curves'][amplitude] = self.plots[module]['amp_plot'].plot(
-                    pen=pg.mkPen(color, width=LINE_WIDTH), name=f"Amp: {amplitude}")
-                self.plots[module]['phase_curves'][amplitude] = self.plots[module]['phase_plot'].plot(
-                    pen=pg.mkPen(color, width=LINE_WIDTH), name=f"Amp: {amplitude}")
-            
-            # Update the curves
-            self.plots[module]['amp_curves'][amplitude].setData(freqs, converted_amps)
-            self.plots[module]['phase_curves'][amplitude].setData(freqs, phases)
-            
-            # If this is a new curve, update legends for proper unit display
-            if is_new_curve:
-                self._update_legends_for_unit_mode()
-
-    def update_data(self, module: int, freqs: np.ndarray, amps: np.ndarray, phases: np.ndarray):
-        """Update the plot data for a specific module."""
-        # Store the raw data for unit conversion
-        iq_data = amps * np.exp(1j * np.radians(phases))  # Reconstruct complex data
-        
-        # Initialize dictionaries if needed
-        if module not in self.raw_data:
-            self.raw_data[module] = {}
-        if module not in self.data:
-            self.data[module] = {}
-        
-        # Store under 'default' key
-        self.raw_data[module]['default'] = (freqs, amps, phases, iq_data)
-        self.data[module]['default'] = (freqs, amps, phases)
-        
-        if module in self.plots:
-            # Only show default curve if no amplitude-specific curves exist yet
-            if len(self.plots[module]['amp_curves']) == 0:
-                # Convert amplitude to selected units
-                converted_amps = UnitConverter.convert_amplitude(
-                    amps, iq_data, self.unit_mode, normalize=self.normalize_magnitudes)
-                
-                # Convert frequency to GHz for display
-                freq_ghz = freqs
-                
-                # Update plots
-                self.plots[module]['amp_curve'].setData(freq_ghz, converted_amps)
-                self.plots[module]['phase_curve'].setData(freq_ghz, phases)
-    
-    def update_progress(self, module: int, progress: float):
-        """Update the progress bar for a specific module."""
-        if module in self.progress_bars:
-            self.progress_bars[module].setValue(int(progress))
-    
-    def complete_analysis(self, module: int):
-        """Mark analysis as complete for a module."""
-        if module in self.progress_bars:
-            self.progress_bars[module].setValue(100)
-            # Check if all modules are complete to hide progress bars
-            self._check_all_complete()
-    
-    def _export_data(self):
-        """Export the collected data with all unit conversions and metadata."""
-        if not self.data:
-            QtWidgets.QMessageBox.warning(self, "No Data", "No data to export yet.")
-            return
-        
-        dialog = QtWidgets.QFileDialog(self)
-        dialog.setAcceptMode(QtWidgets.QFileDialog.AcceptMode.AcceptSave)
-        dialog.setNameFilters([
-            "Pickle Files (*.pkl)",
-            "CSV Files (*.csv)",
-            "All Files (*)"
-        ])
-        dialog.setDefaultSuffix("pkl")
-        
-        if dialog.exec():
-            filename = dialog.selectedFiles()[0]
-            
-            try:
-                if filename.endswith('.pkl'):
-                    self._export_to_pickle(filename)
-                elif filename.endswith('.csv'):
-                    self._export_to_csv(filename)
-                else:
-                    # Default to pickle with same comprehensive format
-                    self._export_to_pickle(filename)
-                    
-                QtWidgets.QMessageBox.information(self, "Export Complete", 
-                                                f"Data exported to {filename}")
-                
-            except Exception as e:
-                traceback.print_exc()  # Print stack trace to console
-                QtWidgets.QMessageBox.critical(self, "Export Error", 
-                                            f"Error exporting data: {str(e)}")
-    
-    def _export_to_pickle(self, filename):
-        """Export data to a pickle file."""
-        # Create comprehensive data structure with all units and metadata
-        export_data = {
-            'timestamp': datetime.datetime.now().isoformat(),
-            'parameters': self.current_params.copy() if hasattr(self, 'current_params') else {},
-            'modules': {}
-        }
-        
-        # Process each module's data
-        for module, data_dict in self.raw_data.items():
-            export_data['modules'][module] = {}
-            
-            # Track measurement index for each module
-            meas_idx = 0
-            
-            for key, data_tuple in data_dict.items():
-                # Extract data and amplitude
-                amplitude, freqs, amps, phases, iq_data = self._extract_data_for_export(key, data_tuple)
-                
-                # Convert to all unit types
-                counts = amps  # Already in counts
-                volts = UnitConverter.convert_amplitude(amps, iq_data, unit_mode="volts")
-                dbm = UnitConverter.convert_amplitude(amps, iq_data, unit_mode="dbm")
-                
-                # Also include normalized versions
-                counts_norm = UnitConverter.convert_amplitude(amps, iq_data, unit_mode="counts", normalize=True)
-                volts_norm = UnitConverter.convert_amplitude(amps, iq_data, unit_mode="volts", normalize=True)
-                dbm_norm = UnitConverter.convert_amplitude(amps, iq_data, unit_mode="dbm", normalize=True)
-                
-                # Use iteration number as key instead of formatted amplitude
-                export_data['modules'][module][meas_idx] = {
-                    'sweep_amplitude': amplitude,
-                    'frequency': {
-                        'values': freqs.tolist(),
-                        'unit': 'Hz'
-                    },
-                    'magnitude': {
-                        'counts': {
-                            'raw': counts.tolist(),
-                            'normalized': counts_norm.tolist(),
-                            'unit': 'counts'
-                        },
-                        'volts': {
-                            'raw': volts.tolist(),
-                            'normalized': volts_norm.tolist(),
-                            'unit': 'V'
-                        },
-                        'dbm': {
-                            'raw': dbm.tolist(),
-                            'normalized': dbm_norm.tolist(),
-                            'unit': 'dBm'
-                        }
-                    },
-                    'phase': {
-                        'values': phases.tolist(),
-                        'unit': 'degrees'
-                    },
-                    'complex': {
-                        'real': iq_data.real.tolist(),
-                        'imag': iq_data.imag.tolist()
-                    }
-                }
-                meas_idx += 1
-
-            export_data['modules'][module]['resonances_hz'] = self.resonance_freqs.get(module, [])
-        
-        # Save the enhanced data structure
-        with open(filename, 'wb') as f:
-            pickle.dump(export_data, f)
-    
-    def _export_to_csv(self, filename):
-        """Export data to CSV files."""
-        base, ext = os.path.splitext(filename)
-        
-        # First create a metadata CSV with the parameters
-        meta_filename = f"{base}_metadata{ext}"
-        with open(meta_filename, 'w', newline='') as f:
-            writer = csv.writer(f)
-            writer.writerow(['Parameter', 'Value'])
-            writer.writerow(['Export Date', datetime.datetime.now().isoformat()])
-            
-            # Add all parameters
-            if hasattr(self, 'current_params'):
-                writer.writerow(['', ''])
-                writer.writerow(['Measurement Parameters', ''])
-                for param, value in self.current_params.items():
-                    # Convert Hz to MHz for frequency parameters
-                    if param in ['fmin', 'fmax', 'max_span'] and isinstance(value, (int, float)):
-                        writer.writerow([param, f"{value/1e6} MHz"])
-                    else:
-                        writer.writerow([param, value])
-
-            if self.resonance_freqs:
-                writer.writerow(['', ''])
-                writer.writerow(['Resonances (Hz)', ''])
-                for module, freqs in self.resonance_freqs.items():
-                    freq_str = ','.join(str(f) for f in freqs)
-                    writer.writerow([f'Module {module}', freq_str])
-        
-        # Now export each module's data
-        for module, data_dict in self.raw_data.items():
-            idx = 0  # Track measurement index
-            for key, data_tuple in data_dict.items():
-                # Extract data and amplitude
-                amplitude, freqs, amps, phases, iq_data = self._extract_data_for_export(key, data_tuple)
-                
-                # Export for each unit type
-                for unit_mode in ["counts", "volts", "dbm"]:
-                    converted_amps = UnitConverter.convert_amplitude(amps, iq_data, unit_mode=unit_mode)
-                    
-                    unit_label = unit_mode
-                    if unit_mode == "dbm":
-                        unit_label = "dBm"
-                    elif unit_mode == "volts":
-                        unit_label = "V"
-                    
-                    csv_filename = f"{base}_module{module}_idx{idx}_{unit_mode}{ext}"
-                    with open(csv_filename, 'w', newline='') as f:
-                        writer = csv.writer(f)
-                        writer.writerow(['# Amplitude:', f"{amplitude}" if 'amplitude' in locals() else "Unknown"])
-                        if unit_mode == "dbm":
-                            writer.writerow(['Frequency (Hz)', f'Power ({unit_label})', 'Phase (deg)'])
-                        else:
-                            writer.writerow(['Frequency (Hz)', f'Amplitude ({unit_label})', 'Phase (deg)'])
-                        
-                        for freq, amp, phase in zip(freqs, converted_amps, phases):
-                            writer.writerow([freq, amp, phase])
-                idx += 1
-    
-    def _extract_data_for_export(self, key, data_tuple):
-        """Extract and prepare data for export from a data tuple."""
-        if key != 'default':
-            if len(data_tuple) >= 5:  # New format with amplitude included
-                freqs, amps, phases, iq_data, amplitude = data_tuple
-            else:
-                # Try to extract amplitude from key format "module_amp"
-                freqs, amps, phases, iq_data = data_tuple
-                try:
-                    amplitude = float(key.split('_')[-1])
-                except (ValueError, IndexError):
-                    amplitude = DEFAULT_AMPLITUDE
-        else:
-            amplitude = DEFAULT_AMPLITUDE  # Default for non-amplitude-specific data
-            freqs, amps, phases, iq_data = data_tuple
-            
-        return amplitude, freqs, amps, phases, iq_data
-
-    def _unwrap_cable_delay_action(self):
-        """
-        Fits the phase data of the first curve in the active module's plot,
-        calculates the corresponding cable length, updates the phase curves,
-        and adjusts the cable length spinner.
-        """
-        if not self.raw_data:
-            QtWidgets.QMessageBox.information(self, "No Data", "No network analysis data available to process.")
-            return
-
-        current_tab_index = self.tabs.currentIndex()
-        if current_tab_index < 0:
-            QtWidgets.QMessageBox.warning(self, "No Module Selected", "Please select a module tab.")
-            return
-        
-        active_module_text = self.tabs.tabText(current_tab_index)
-        try:
-            active_module = int(active_module_text.split(" ")[1])
-        except (IndexError, ValueError):
-            QtWidgets.QMessageBox.critical(self, "Error", f"Could not determine active module from tab: {active_module_text}")
-            return
-
-        if active_module not in self.raw_data or not self.raw_data[active_module]:
-            QtWidgets.QMessageBox.information(self, "No Data", f"No data for Module {active_module}.")
-            return
-        
-        module_data_dict = self.raw_data[active_module]
-        
-        target_key = None
-        if 'amps' in self.original_params and self.original_params['amps']:
-            first_amplitude_setting = self.original_params['amps'][0]
-            potential_key = f"{active_module}_{first_amplitude_setting}"
-            if potential_key in module_data_dict:
-                target_key = potential_key
-        
-        if target_key is None and 'default' in module_data_dict:
-            target_key = 'default'
-        
-        if target_key is None:
-            if module_data_dict:
-                target_key = next(iter(module_data_dict))
-            else:
-                QtWidgets.QMessageBox.information(self, "No Data", f"No sweep data found for Module {active_module} to process.")
-                return
-
-        # --- Extract data for the chosen key ---
-        # data_tuple: (freqs, amps, phases_displayed_deg, iq_data, amplitude_setting) OR (freqs, amps, phases_displayed_deg, iq_data)
-        data_tuple = module_data_dict[target_key]
-        
-        if len(data_tuple) == 5:
-            freqs_active, _, phases_displayed_active_deg, _, _ = data_tuple
-        elif len(data_tuple) == 4:
-            freqs_active, _, phases_displayed_active_deg, _ = data_tuple
-        else:
-            QtWidgets.QMessageBox.critical(self, "Error", "Unexpected data format for selected sweep.")
-            return
-
-        if len(freqs_active) == 0:
-            QtWidgets.QMessageBox.information(self, "No Data", "Selected sweep has no frequency data.")
-            return
-
-        # --- Determine the old cable length used for this specific sweep ---
-        # This is tricky. For now, assume current_params['cable_length'] was used for the "first" sweep.
-        # A more robust solution would store the cable_length with each sweep in self.raw_data.
-        L_old_physical = self.current_params.get('cable_length', DEFAULT_CABLE_LENGTH)
-
-        # --- Perform Fit and Calculation ---
-        try:
-            tau_additional = fit_cable_delay(freqs_active, phases_displayed_active_deg)
-            L_new_physical = calculate_new_cable_length(L_old_physical, tau_additional)
-        except Exception as e:
-            QtWidgets.QMessageBox.critical(self, "Calculation Error", f"Error during cable delay calculation: {str(e)}")
-            traceback.print_exc()
-            return
-
-        # --- Update All Phase Curves for the Active Module ---
-        if active_module in self.plots:
-            plot_info = self.plots[active_module]
-            
-            # Update amplitude-specific phase curves
-            for amp_key_iter, curve_item in plot_info['phase_curves'].items():
-                # amp_key_iter is the amplitude value (float)
-                # We need to find the corresponding full key in raw_data to get original displayed phases
-                raw_data_key_for_curve = f"{active_module}_{amp_key_iter}"
-                if raw_data_key_for_curve in module_data_dict:
-                    data_tuple_curve = module_data_dict[raw_data_key_for_curve]
-                    if len(data_tuple_curve) == 5:
-                        freqs_curve, _, phases_deg_current_display_curve, _, _ = data_tuple_curve
-                    elif len(data_tuple_curve) == 4: # Should not happen for amp_specific curves but handle
-                        freqs_curve, _, phases_deg_current_display_curve, _ = data_tuple_curve
-                    else:
-                        continue # Skip malformed data
-
-                    if len(freqs_curve) > 0:
-                        # Assume this curve was also compensated with L_old_physical.
-                        # This is an approximation if the user changed cable length between sweeps
-                        # without a full re-run.
-                        # First get the unwrapped phase data
-                        new_phases_deg_for_curve = recalculate_displayed_phase(
-                            freqs_curve, phases_deg_current_display_curve, L_old_physical, L_new_physical
-                        )
-                        
-                        # Recenter the unwrapped phase data so the first point begins at 0 phase
-                        if len(new_phases_deg_for_curve) > 0:
-                            first_point_phase = new_phases_deg_for_curve[0]
-                            new_phases_deg_for_curve = new_phases_deg_for_curve - first_point_phase
-                            
-                        # Now re-wrap phases to [-180, 180] range
-                        new_phases_deg_for_curve = ((new_phases_deg_for_curve + 180) % 360) - 180
-                            
-                        curve_item.setData(freqs_curve, new_phases_deg_for_curve)
-            
-            # Update the main/default phase curve (if it holds data)
-            # The main curve 'phase_curve' should only have data if no amp-specific curves exist.
-            if not plot_info['phase_curves'] and 'default' in module_data_dict:
-                main_curve_item = plot_info['phase_curve']
-                data_tuple_main = module_data_dict['default']
-                if len(data_tuple_main) == 4:
-                    freqs_main, _, phases_deg_current_display_main, _ = data_tuple_main
-                    if len(freqs_main) > 0:
-                        # First get the unwrapped phase data
-                        new_phases_deg_for_main = recalculate_displayed_phase(
-                            freqs_main, phases_deg_current_display_main, L_old_physical, L_new_physical
-                        )
-                        
-                        # Recenter the unwrapped phase data so the first point begins at 0 phase
-                        if len(new_phases_deg_for_main) > 0:
-                            first_point_phase = new_phases_deg_for_main[0]
-                            new_phases_deg_for_main = new_phases_deg_for_main - first_point_phase
-                            
-                        # Now re-wrap phases to [-180, 180] range
-                        new_phases_deg_for_main = ((new_phases_deg_for_main + 180) % 360) - 180
-                        
-                        main_curve_item.setData(freqs_main, new_phases_deg_for_main)
-
-            plot_info['phase_plot'].enableAutoRange(pg.ViewBox.YAxis, True)
-
-
-        # --- Update Cable Length Spinner and Stored Parameters ---
-        self.cable_length_spin.setValue(L_new_physical)
-        self.current_params['cable_length'] = L_new_physical
-        
-        QtWidgets.QMessageBox.information(self, "Cable Delay Updated", 
-                                          f"Cable length updated to {L_new_physical:.3f} m based on phase fit.")
-
-class InitializeCRSDialog(QtWidgets.QDialog):
-    """Dialog for CRS initialization options."""
-    def __init__(self, parent=None, crs_obj=None):
-        super().__init__(parent)
-        self.crs = crs_obj 
-        self.setWindowTitle("Initialize CRS Board")
-        self.setModal(True)
-        
-        layout = QtWidgets.QVBoxLayout(self)
-        
-        # IRIG Time Source Group
-        irig_group = QtWidgets.QGroupBox("IRIG Time Source")
-        irig_layout = QtWidgets.QVBoxLayout(irig_group)
-        
-        self.rb_backplane = QtWidgets.QRadioButton("BACKPLANE")
-        self.rb_test = QtWidgets.QRadioButton("TEST")
-        self.rb_sma = QtWidgets.QRadioButton("SMA")
-        
-        # Default selection (e.g., TEST)
-        self.rb_test.setChecked(True) 
-        
-        irig_layout.addWidget(self.rb_backplane)
-        irig_layout.addWidget(self.rb_test)
-        irig_layout.addWidget(self.rb_sma)
-        layout.addWidget(irig_group)
-        
-        # Clear Channels Checkbox
-        self.cb_clear_channels = QtWidgets.QCheckBox("Clear all channels on this module")
-        self.cb_clear_channels.setChecked(True) # Default to checked
-        layout.addWidget(self.cb_clear_channels)
-        
-        # Buttons
-        btn_layout = QtWidgets.QHBoxLayout()
-        self.ok_btn = QtWidgets.QPushButton("OK")
-        self.cancel_btn = QtWidgets.QPushButton("Cancel")
-        btn_layout.addStretch()
-        btn_layout.addWidget(self.ok_btn)
-        btn_layout.addWidget(self.cancel_btn)
-        
-        layout.addLayout(btn_layout)
-        
-        self.ok_btn.clicked.connect(self.accept)
-        self.cancel_btn.clicked.connect(self.reject)
-
-    def get_selected_irig_source(self):
-        if self.crs is None: # Should not happen if button is disabled correctly
-            return None 
-            
-        if self.rb_backplane.isChecked():
-            return self.crs.TIMESTAMP_PORT.BACKPLANE
-        elif self.rb_test.isChecked():
-            return self.crs.TIMESTAMP_PORT.TEST
-        elif self.rb_sma.isChecked():
-            return self.crs.TIMESTAMP_PORT.SMA
-        return None # Should have one selected
-
-    def get_clear_channels_state(self) -> bool:
-        return self.cb_clear_channels.isChecked()
-
-# ───────────────────────── Find Resonances Dialog ─────────────────────────
-class FindResonancesDialog(QtWidgets.QDialog):
-    """Dialog for configuring parameters for fitting.find_resonances."""
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Find Resonances Parameters")
-        self.setModal(True)
-        self._setup_ui()
-
-    def _setup_ui(self):
-        layout = QtWidgets.QFormLayout(self)
-
-        # expected_resonances (int or None)
-        self.expected_resonances_edit = QtWidgets.QLineEdit()
-        self.expected_resonances_edit.setPlaceholderText("Optional (e.g., 10)")
-        # Validator for integer or empty string
-        regex_int_or_empty = QRegularExpression("^$|^[1-9][0-9]*$") # Empty or positive integer
-        self.expected_resonances_edit.setValidator(QRegularExpressionValidator(regex_int_or_empty, self))
-        layout.addRow("Expected Resonances:", self.expected_resonances_edit)
-
-        # min_dip_depth_db (float)
-        self.min_dip_depth_db_edit = QtWidgets.QLineEdit(str(1.0)) # Default from find_resonances
-        self.min_dip_depth_db_edit.setValidator(QDoubleValidator(0.01, 100.0, 2, self)) # Min, Max, Decimals
-        layout.addRow("Min Dip Depth (dB):", self.min_dip_depth_db_edit)
-
-        # min_Q (float)
-        self.min_Q_edit = QtWidgets.QLineEdit(str(1e4)) # Default
-        self.min_Q_edit.setValidator(QDoubleValidator(1.0, 1e9, 0, self))
-        layout.addRow("Min Q:", self.min_Q_edit)
-
-        # max_Q (float)
-        self.max_Q_edit = QtWidgets.QLineEdit(str(1e7)) # Default
-        self.max_Q_edit.setValidator(QDoubleValidator(1.0, 1e9, 0, self))
-        layout.addRow("Max Q:", self.max_Q_edit)
-
-        # min_resonance_separation_mhz (float, will be converted to Hz)
-        self.min_resonance_separation_mhz_edit = QtWidgets.QLineEdit(str(0.1)) # Default 100e3 Hz = 0.1 MHz
-        self.min_resonance_separation_mhz_edit.setValidator(QDoubleValidator(0.001, 1000.0, 3, self))
-        layout.addRow("Min Separation (MHz):", self.min_resonance_separation_mhz_edit)
-        
-        # data_exponent (float)
-        self.data_exponent_edit = QtWidgets.QLineEdit(str(2.0)) # Default
-        self.data_exponent_edit.setValidator(QDoubleValidator(0.1, 10.0, 2, self))
-        layout.addRow("Data Exponent:", self.data_exponent_edit)
-
-        # Buttons
-        self.button_box = QtWidgets.QDialogButtonBox(
-            QtWidgets.QDialogButtonBox.StandardButton.Ok | QtWidgets.QDialogButtonBox.StandardButton.Cancel
-        )
-        self.button_box.accepted.connect(self.accept)
-        self.button_box.rejected.connect(self.reject)
-        layout.addRow(self.button_box)
-
-    def get_parameters(self) -> dict | None:
-        params = {}
-        try:
-            # Expected Resonances (optional int)
-            expected_text = self.expected_resonances_edit.text().strip()
-            if expected_text:
-                params['expected_resonances'] = int(expected_text)
-            else:
-                params['expected_resonances'] = None
-
-            params['min_dip_depth_db'] = float(self.min_dip_depth_db_edit.text())
-            params['min_Q'] = float(self.min_Q_edit.text())
-            params['max_Q'] = float(self.max_Q_edit.text())
-            
-            min_sep_mhz = float(self.min_resonance_separation_mhz_edit.text())
-            params['min_resonance_separation_hz'] = min_sep_mhz * 1e6 # Convert MHz to Hz
-            
-            params['data_exponent'] = float(self.data_exponent_edit.text())
-            
-            # Basic validation
-            if params['min_Q'] >= params['max_Q']:
-                QtWidgets.QMessageBox.warning(self, "Validation Error", "Min Q must be less than Max Q.")
-                return None
-            if params['min_dip_depth_db'] <=0:
-                QtWidgets.QMessageBox.warning(self, "Validation Error", "Min Dip Depth must be positive.")
-                return None
-
-            return params
-        except ValueError as e:
-            QtWidgets.QMessageBox.critical(self, "Input Error", f"Invalid input: {str(e)}")
-            return None
-
-# ───────────────────────── Main Application ─────────────────────────
 class Periscope(QtWidgets.QMainWindow):
     """
     Multi‑pane PyQt application for real-time data visualization of:
